% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{add_celltype_info}
\alias{add_celltype_info}
\title{Add flytable cell type information to a dataframe with flywire ids}
\usage{
add_celltype_info(x, ...)
}
\arguments{
\item{x}{a data.frame containing root ids}

\item{...}{additional arguments passed to \code{flytable_cell_types}}
}
\value{
a data.frame with extra columns
}
\description{
Add flytable cell type information to a dataframe with flywire ids
}
\details{
the root ids must be in a column called one of \code{"pre_id",
  "post_id", "root_id"}
}
\examples{
\donttest{
kcin=flywire_partner_summary("720575940626474889", partners = 'in',
  cleft.threshold = 50)
kcin
kcin2=add_celltype_info(kcin)
kcin2
library(dplyr)
kcin2 \%>\%
  group_by(cell_type) \%>\%
  summarise(wt = sum(weight),n=n()) \%>\%
  arrange(desc(wt))
kcin2 \%>\%
  count(cell_class, wt = weight)
}
}
\seealso{
\code{\link{flytable_cell_types}}
}
