% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng-mesh.R
\name{as.mesh3d.ng_raw}
\alias{as.mesh3d.ng_raw}
\alias{as.mesh3d.ng_raw_list}
\title{Convert a raw neuroglancer mesh into a mesh3d object}
\usage{
\method{as.mesh3d}{ng_raw}(x, ...)

\method{as.mesh3d}{ng_raw_list}(x, ...)
}
\arguments{
\item{x}{An ng_raw or ng_raw_list object}

\item{...}{additional arguments passed to \code{\link{tmesh3d}}}
}
\description{
Convert a raw neuroglancer mesh into a mesh3d object
}
