% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincircuits.R
\name{braincircuits_register}
\alias{braincircuits_register}
\alias{braincircuits_login}
\alias{braincircuits_token}
\title{Create and log into a braincircuits account}
\usage{
braincircuits_register(open = TRUE, email = NULL, password = NULL)

braincircuits_login(email = NULL, password = NULL)

braincircuits_token(email = NULL, password = NULL)
}
\arguments{
\item{open}{logical, whether or not to open a browser window to the \href{https://braincircuits.io/app/login}{login page} for \href{https://braincircuits.io/}{braincircuits.io}.}

\item{email}{the email you have registered/want to register with \href{https://braincircuits.io/}{braincircuits.io}.}

\item{password}{the password you want to set (when using \code{braincircuits_register}) or have set (when using \code{braincircuits_login}) for your account.}
}
\value{
A bearer access token with which to query the braincircuits API.
}
\description{
\href{https://braincircuits.io/}{braincircuits} is a website built by Stephan Gerhard that provides some services for FAFB flywire data. For example, information on predictions for dense core vesicles (DCVs).
These functions assist you in registering and logging into a braincircuits account in order to query its API. You must first log into the website \href{https://braincircuits.io/app/login}{here}.
You then run \code{braincircuits_login}, which will call \code{braincircuits_register}. At the moment, this function is only needed for \code{flywire_dcvs} and is called under the hood there. Therefore, beyond testing that you
can connect to \href{https://braincircuits.io/}{braincircuits.io} there is little reason to call this function directly.
}
\examples{
\donttest{
# Run this function:
braincircuits_login()
# If you run it for the first time, a window will open
# to the braincircuits log in page. Log in with an email.
# then when prompted in the R command line, re-give your email address and
# give a new password. You can save these in your .Renviron file to prevent having
# to do this in future.
}
}
\seealso{
\code{\link{flywire_dcvs}}
}
