% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_fetch}
\alias{brainmaps_fetch}
\title{GET/POST from brainmaps API}
\usage{
brainmaps_fetch(url, body = NULL, parse.json = TRUE,
  include_headers = FALSE, simplifyVector = TRUE, ...)
}
\value{
An R list parse
}
\description{
GET/POST from brainmaps API
}
\examples{
\dontrun{
brainmaps_fetch("https://brainmaps.googleapis.com/v1/volumes")
}
}
