% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_listfragments}
\alias{brainmaps_listfragments}
\title{Low level call to brainmaps API to list mesh fragment ids for segment ids}
\usage{
brainmaps_listfragments(x,
  volume = "772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2",
  meshName = "mcws_quad1e6", ...)
}
\arguments{
\item{x}{Single segment identifier}

\item{volume}{String identifier for the volume containing segments}

\item{meshName}{String identifier for the meshes}

\item{...}{Additional arguments passed to \code{\link{brainmaps_fetch}}}
}
\value{
Character vector of fragment ids
}
\description{
Low level call to brainmaps API to list mesh fragment ids for segment ids
}
\examples{
\dontrun{
brainmaps_listfragments(7186840767)
}
}
\seealso{
\code{\link{read_brainmaps_meshes}}
}
