% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_volume}
\alias{brainmaps_volume}
\title{Extract brainmaps volume identifier}
\usage{
brainmaps_volume(x)
}
\arguments{
\item{x}{character vector containing brainmaps URI, a parsed \code{brainmaps_uri} object or a character vector already containing a volume specifier.}
}
\value{
character vector containing a volume specifier.
}
\description{
Extract brainmaps volume identifier
}
\examples{
brainmaps_volume("772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2")
brainmaps_volume("brainmaps://772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2")
}
