% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_xyz2id}
\alias{brainmaps_xyz2id}
\title{Convert xyz locations in brainmaps volumes to segmentation ids}
\usage{
brainmaps_xyz2id(xyz,
  volume = "772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2",
  voxdims = c(8, 8, 40), ...)
}
\arguments{
\item{xyz}{Nx3 matrix of points or an object containing vertex data that is
compatible with \code{\link{xyzmatrix}}. These should be in physical space
(i.e. nm) unless \code{voxdims=NULL}.}

\item{volume}{character vector identifier string for the volume containing
segmentation data - see examples}

\item{voxdims}{the implied voxel dimensions for the volume. If set to
\code{NULL} then the function will not attempt to scale the incoming xyz
locations.}

\item{...}{Additional arguments passed to \code{\link{brainmaps_fetch}}}
}
\value{
A numeric vector of Google segment ids
}
\description{
Convert xyz locations in brainmaps volumes to segmentation ids
}
