% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid-ngl.R
\name{catmaid2ngl}
\alias{catmaid2ngl}
\alias{catmaid2ngl.neuron}
\alias{catmaid2ngl.neuronlist}
\alias{catmaid2ngl.character}
\alias{catmaid2ngl.default}
\title{Convert CATMAID things (eg neurons) into Neuroglancer equivalents}
\usage{
catmaid2ngl(x, ...)

\method{catmaid2ngl}{neuron}(x,
  chunksize = getOption("fafbseg.brainmaps_xyz2id.chunksize", 4000), ...)

\method{catmaid2ngl}{neuronlist}(x, OmitFailures = TRUE, ...)

\method{catmaid2ngl}{character}(x, open = FALSE, ...)

\method{catmaid2ngl}{default}(x, ...)
}
\arguments{
\item{x}{An object to convert (see Descriptions for each method)}

\item{...}{Additional arguments passed to methods}

\item{chunksize}{send queries in batches each of which has at most
\code{chunksize} points. The default is chosen since the brainmaps API can
time out if the points take too long to map (more likely if they are spread
out across the brain). There is also a maximum number of points per call
(10,000 was the recommended upper limit at one point).}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in nlapply stopping with
an error message the moment there is an error. For other values, see
details.}

\item{open}{Whether or not to open the URL in a browser - this defaults to
\code{TRUE} in interactive use.}
}
\description{
\code{catmaid2ngl.neuron} uses \code{\link{brainmaps_xyz2id}}
  and \code{\link{read_segments2}} to find the skeletons corresponding to a
  (catmaid) neuron

\code{catmaid2ngl.neuronlist} applies \code{catmaid2ngl.neuron}
  over a \code{\link{neuronlist}}

\code{catmaid2ngl.character} uses \code{\link{open_fafb_ngl}} to
  open CATMAID at a location matching a Neuroglancer URL.

Converts a CATMAID skid specification (see
  \code{\link{catmaid_skids}}) for one neuron into a an autosegmentation
  based neuron using \code{catmaid2ngl.neuron}.
}
