% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose-segmentation.R
\name{choose_segmentation}
\alias{choose_segmentation}
\alias{with_segmentation}
\title{Choose or (temporarily) use a FAFB autosegmentation}
\usage{
choose_segmentation(
  release = c("20190805", "20190521", "flywire31"),
  set = TRUE
)

with_segmentation(release, expr)
}
\arguments{
\item{release}{character vector specifying a released segmentation.}

\item{set}{Whether or not to set the selected options for the selected
\code{release}.}

\item{expr}{The expression to execute with the temporary options set}
}
\value{
If \code{set=TRUE} a list containing the previous values of the
  relevant global options (in the style of \code{\link{options}}. If
  \code{set=FALSE} a named list containing the option values.

The result of evaluating \code{expr}
}
\description{
\code{with_segmentation} allows a specific segmentation to be
  temporarily selected.
}
\details{
Each released segmentation implies a number of global options.
}
\examples{
\donttest{
choose_segmentation('20190805', set=FALSE)
}
\donttest{
n <- with_segmentation("20190521",{
  read.neuron.brainmaps(22427007374)
})

\dontrun{
# open location in flywire
with_segmentation("flywire", open_fafb_ngl(c(433440, 168344, 131200)))
}
}
}
