% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compare_ng_neuron}
\alias{compare_ng_neuron}
\title{Compare a neuroglancer mesh object with a regular neuron}
\usage{
compare_ng_neuron(x, n, breaks = 3, colpal = c("cyan", "red"),
  plot = TRUE, plotn = plot, ...)
}
\arguments{
\item{x}{A neuroglancer mesh, rgl::mesh3d object or other object with a
defined xyzmatrix function.}

\item{n}{A nat::neuron object}

\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{colpal}{A function defining a colour palette or a vector of colour
names. Should}

\item{plot}{Whether to plot anything (set to \code{FALSE} when you just want to
get the distance information)}

\item{plotn}{Whether to plot the neuron \code{n}}

\item{...}{Additional arguments passed to \code{plot3d.neuron}}
}
\value{
Invisibly, a data.frame with the distances of each object in \code{x}
  to its nearest neighbour in \code{n} as well as the breaks used for
  colouring points when plotting.
}
\description{
Compare a neuroglancer mesh object with a regular neuron
}
\examples{
\dontrun{
x=read_ng_raw('meshdata/')
library(elmr)
y=read.neuron.catmaid(23432)
compare_ng_neuron(x,y)
}
}
