% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dr_fafbseg}
\alias{dr_fafbseg}
\title{Print information about fafbseg setup including tokens and python modules}
\usage{
dr_fafbseg(pypkgs = NULL)
}
\arguments{
\item{pypkgs}{Additional python packages to check beyond the standard ones
that \bold{fafbseg} knows about such as \code{cloudvolume}. When set to
\code{FALSE}, this turns off the Python package report altogether.}
}
\description{
Print information about your \bold{fafbseg} setup including your
  FlyWire/ChunkedGraph authentication tokens, Python modules and the
  nat.h5reg / java setup required for transforming points between EM and
  light level template brains.
}
\examples{
\dontrun{
dr_fafbseg(pymodules=FALSE)
}
}
