% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire.R
\name{fafb2flywire}
\alias{fafb2flywire}
\title{Map points in FAFB14 space (xyz nm) to FlyWire v1}
\usage{
fafb2flywire(xyz, method = c("mapmany", "map1"), chunksize = 200, ...)
}
\arguments{
\item{xyz}{A Nx3 matrix of points}

\item{method}{Whether to map many points at once (default) or just one}

\item{...}{Additional arguments passed to httr::GET/POST operation}
}
\value{
an Nx3 matrix of points
}
\description{
Map points in FAFB14 space (xyz nm) to FlyWire v1
}
\examples{
fafb142flywirev1(cbind(163826*4, 75263*4, 3513*40))
fafb142flywirev1(cbind(163826*4, 75263*4, 3513*40), method="map1")

data("AV4b1")
set.seed(42)
before=xyzmatrix(AV4b1)[sample(nvertices(AV4b1), size=2000), ]
after=fafb142flywirev1(before)
d=sqrt(rowSums((before-after)^2))
hist(d, br=20)
}
