% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg-package.R
\docType{package}
\name{fafbseg-package}
\alias{fafbseg}
\alias{fafbseg-package}
\title{fafbseg: Support Functions for Analysis of FAFB EM Segmentation}
\description{
Functions that read some of the raw data formats produced by Google electron microscopy segmentation tools as displayed by 'neuroglancer' and then allow basic 3D visualisation and analysis.
}
\section{Package Options}{
 \itemize{

  \item{\code{fafbseg.sampleurl}} optionally set to a sample Neuroglancer URL
  that will modified to point to arbitrary locations by
  \code{\link{open_fafb_ngl}}.

  \item{\code{fafbseg.cave.datastack_name}} optionally set to the
  \code{datastack_name} used by the \code{CAVE} system. Expert use only.
  Normally \code{\link{choose_segmentation}} looks after this. See
  \code{\link{flywire_cave_query}} for further details.

  \item{\code{fafbseg.caveclient.loglevel}} The caveclient logger level is
  set to \code{"WARNING"} on initialisation. Genuine caveclient \code{WARNING}
  / \code{ERROR} records still propagate but \code{INFO} is suppressed.
  Override \code{"INFO"} (or any other Python \code{logging} level name). See
  \code{\link{flywire_cave_client}} for additional details.

  \item{\code{fafbseg.condaenv}} the python environment to use with
  miniconda. Defaults to \code{"r-reticulate"}, which is the environment set
  up by \code{reticulate::install_miniconda}. You can set this to something
  different if you would like to keep a separate miniconda virtual
  environment just for fafbseg (expert use only).

  \item{\code{fafbseg.sqlitepath}} optional to set the location of SQLite
  tables used by \code{\link{flywire_partners}} and friends.

  \item{\code{fafbseg.cachedir}} The location for disk caches for functions
  including \code{\link{flywire_leaves}} If unset on package load, will be
  set to an appropriate user folder using
  \code{rappdirs::\link[rappdirs]{user_data_dir}}.

  \item{\code{fafbseg.flcachesize}} The maximum cache size in bytes. Defaults
  to \code{1.5 * 1024^3} when unset. See \code{\link{flywire_leaves}} for
  details.

  \item{\code{fafbseg.flywire_roots.chunksize}} this will default to
  processing 100,000 rootids at a time. Set this smaller if the queries time
  out / give "Request Entity Too Large" errors. Larger settings are unlikely
  to have much of a speed impact. See \code{\link{flywire_rootid}} for
  details.

  \item{\code{fafbseg.skelziproot}} set to the location of a folder
  containing the zipped versions of the skeletonised segmentations. This will
  be used by \code{\link{read_segments}}, \code{\link{read_topn}} etc.

  \item{\code{fafbseg.skeletonuri}} a brainmaps URI specifying a remote
  source used by \code{\link{read.neurons.brainmaps}} and
  \code{\link{brainmaps_skeleton}} to read neuronal skeletons.

  \item{\code{fafbseg.brainmaps_xyz2id.chunksize}} this will default to
  querying 4000 vertices at a time. Set this smaller if the queries time out
  or larger to speed things up. See \code{\link{brainmaps_xyz2id}} for
  details.

  \item{\code{fafbseg.flytable.url}} the seatable server URL used by
  \code{\link{flytable_login}}, \code{\link{flytable_base}} and downstream
  helpers including \code{\link{cam_meta}} and
  \code{\link{flytable_cached_table}}. Defaults to
  \code{"https://flytable.mrc-lmb.cam.ac.uk/"}. Setting this option (e.g.
  inside a \code{withr::with_options()} block) lets downstream packages point
  the flytable infrastructure at an alternative seatable instance without
  having to thread the URL through every call site.

  }
}

\examples{
options()[grep("^fafbseg\\\\.", names(options()))]
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/fafbseg}
  \item Report bugs at \url{https://github.com/natverse/fafbseg/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

Other contributors:
\itemize{
  \item Sridhar Jagannathan \email{j.sridharrajan@gmail.com} (\href{https://orcid.org/0000-0002-2078-1145}{ORCID}) [contributor]
  \item Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID}) [contributor]
}

}
\keyword{internal}
