% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nginfo.R
\name{fct2nginfo}
\alias{fct2nginfo}
\title{Convert flytable cell type information into a neuroglancer info file}
\usage{
fct2nginfo(f, ids = NULL, version = NULL, sep = "_", gluestr = NULL, ...)
}
\arguments{
\item{f}{Path to a json file. \code{write_nginfo} will create the enclosing
directory if necessary.}

\item{ids}{FlyWire root ids in any form understood by
\code{\link{flywire_ids}}}

\item{version}{Optional numeric CAVE version (see \code{flywire_cave_query}).
The special signalling value of \code{TRUE} uses the current default data
dump as returned by \code{\link{flywire_connectome_data_version}}.}

\item{sep}{The separator used to paste multiple columns of annotations
together.}

\item{gluestr}{Optional string passed to \code{glue::glue} which is
interpreted in the context of the annotation data.frame produced by
\code{\link{flytable_meta}}. This allows arbitrary formatting for}

\item{...}{Additional arguments passed to \code{\link{flytable_meta}} and
then eventually \code{\link{flytable_cell_types}}.}
}
\value{
The path \code{f} invisibly
}
\description{
Convert flytable cell type information into a neuroglancer info file
}
\examples{
\dontrun{
# all neurons in info table
fct2nginfo(f='path/to/flytablev526/info', version=526)
fct2nginfo(f='path/to/hemilineagev526/info', version=526,
  gluestr="{ito_lee_hemilineage}_{toupper(substr(side,1,1))}")
fct2nginfo("MBON\%", 'path/to/mboninfo/info')
}
}
\seealso{
\code{\link{write_nginfo}}
}
