% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-groups.R
\name{find_merged_segments}
\alias{find_merged_segments}
\title{Find all merged segment ids for a given starting google segment id}
\usage{
find_merged_segments(x)
}
\arguments{
\item{x}{a segment id or any other input that can be interpreted by
\code{\link{ngl_segments}}}

\item{mergeinfo}{The merge information data.frame created by
\code{\link{read_mergeinfo}}}
}
\value{
vector of segment ids
}
\description{
Find all merged segment ids for a given starting google segment id
}
\details{
segment ids in \code{ffn16reseg-ms1000_md0.02_c0.6_iou0.7} always
  match one raw segment id in \code{fafb_v14_16nm_v00c_split3xfill2} but may
  map to a large agglomerated merge group.

  This function is slow but does not require any pre-processing of the table
  of merge information.
}
