% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable_cell_types}
\alias{flytable_cell_types}
\title{Fetch (memoised) flywire cell type information from flytable}
\usage{
flytable_cell_types(
  pattern = NULL,
  version = NULL,
  timestamp = NULL,
  target = c("type", "cell_type", "hemibrain_type", "cell_class", "super_class",
    "ito_lee_hemilineage", "all"),
  table = c("info", "optic", "both"),
  transfer_hemibrain_type = c("extra", "none", "all"),
  cache = TRUE,
  use_static = NA
)
}
\arguments{
\item{pattern}{Optional character vector specifying a pattern that cell types
must match in a SQL \code{LIKE} statement executed by
\code{\link{flytable_query}}. The suffix \code{_L} or \code{_R} can be used
to restricted to neurons annotated to the L or R hemisphere. See examples.}

\item{version}{An optional CAVE materialisation version number. See
\code{\link{flywire_cave_query}} for more details. Note also that the
special signalling value of \code{TRUE} implies the latest locally
available connectome dump.}

\item{timestamp}{An optional timestamp as a string or POSIXct, interpreted as
UTC when no timezone is specified.}

\item{target}{A character vector specifying which flytable columns
\code{pattern} should match. The special value of \code{type} means either
\code{cell_type} \emph{or} \code{hemibrain_type} should match. The special
value of \code{all} means to match against any of \code{cell_type,
hemibrain_type, cell_class}.}

\item{table}{Which cell type information tables to use (\code{info} for
brain, \code{optic} for optic lobes or \code{both}).}

\item{transfer_hemibrain_type}{Whether to transfer the \code{hemibrain_type}
column into the \code{cell_type} (default TRUE, see details)}

\item{cache}{Whether to cache the results for 5m (default \code{TRUE} since
the flytable query is is a little expensive)}

\item{use_static}{Whether to use static cell type information (from Schlegel et al)}
}
\value{
The original data.frame left joined to appropriate rows from
  flytable.
}
\description{
Fetch (memoised) flywire cell type information from flytable
}
\details{
when \code{transfer_hemibrain_type=TRUE}, \code{hemibrain_type}
  values will be transferred into the \code{cell_type} column if
  \code{cell_type} is empty.

  It seems that SQL LIKE searches (e.g. containing the \code{\%} symbol) do
  not work for the \code{ito_lee_hemilineage} column. You can still search
  for exact matches or use full regular expression queries (which operate by
  downloading all rows and then filtering on your machine).

  Static cell type information is provided by Schlegel et 2023
}
\examples{
\donttest{
flytable_cell_types("MBON\%")
flytable_cell_types("MBON\%", version=450)
# the latest connectome dump, see flywire_connectome_data_version()
\dontrun{
flytable_cell_types("MBON\%", version=TRUE)
}

# two characters
flytable_cell_types("MBON__")
# at least one character
flytable_cell_types("MBON_\%")
# range
flytable_cell_types("MBON2[0-5]")

# include side specification
flytable_cell_types("DA2_lPN_R")
# only the RHS MBON20
flytable_cell_types("MBON20_R")
# all RHS cells with class MBON
flytable_cell_types("MBON_R", target="cell_class")

# anything with type *OR* class information
cells=flytable_cell_types(target = 'all')
# anything that mentions PN anywhere
pncands=flytable_cell_types('\%PN\%', target = 'all')
}
}
\seealso{
\code{\link{add_celltype_info}}
}
