% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-arrow.R
\name{flywire_adjacency_matrix2}
\alias{flywire_adjacency_matrix2}
\title{Fast adjacency matrices based on flywire connectome dumps}
\usage{
flywire_adjacency_matrix2(
  rootids = NULL,
  inputids = NULL,
  outputids = NULL,
  sparse = TRUE,
  threshold = 0,
  version = NULL,
  Verbose = interactive()
)
}
\arguments{
\item{rootids}{flywire root ids for the bodies to fetch all by all
connectivity information.}

\item{inputids, outputids}{identifiers for input and output bodies (use as an
alternative to \code{rootids})}

\item{sparse}{Whether to return a sparse matrix (default \code{FALSE})}

\item{threshold}{For \code{flywire_partner_summary} only return partners with
greater than this number of connections to the query neuron(s) (default of
0 returns all connections)}

\item{version}{Optional CAVE version. The default value of \code{NULL} uses
the latest data dump available.}

\item{Verbose}{Logical indication whether to print status messages during the
query (default \code{T} when interactive, \code{F} otherwise).}
}
\value{
A sparse matrix (\code{Matrix::dgCMatrix}) or regular \code{matrix}.
}
\description{
Fast adjacency matrices based on flywire connectome dumps
}
\examples{
\dontrun{
dm2pnkc=flywire_adjacency_matrix2(inputids="DM2_lPN_R", outputids="class:Kenyon_Cell_R")
}
}
