% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_change_log}
\alias{flywire_change_log}
\title{Summarise edits for one or more FlyWire segment ids}
\usage{
flywire_change_log(
  x,
  root_ids = FALSE,
  filtered = TRUE,
  tz = "UTC",
  OmitFailures = TRUE,
  ...
)
}
\arguments{
\item{x}{One or more flywire segment ids (character vector strongly
recommended but any format accepted by \code{\link{ngl_segments}} will
work).}

\item{root_ids}{Whether to look up the root ids before/after each change.
Default \code{FALSE}. NB this is quite resource intensive so please do not
flood the server with requests of this sort.}

\item{filtered}{Whether to filter out edits unlikely to relate to the current
state of the neuron (default \code{TRUE}, see details).}

\item{tz}{Time zone for edit timestamps. Defaults to "UTC" i.e. Universal
Time, Coordinated. Set to "" for your current timezone. See
\code{\link{as.POSIXct}} for more details.}

\item{OmitFailures}{Whether to omit neurons for which there is an API timeout
or error. The default value (\code{TRUE}) will skip over errors, while
\code{NA}) will result in a hard stop on error. See \code{\link{nlapply}}
for more details.}

\item{...}{Additional arguments passed to \code{\link{flywire_fetch}}}
}
\value{
A data frame with values itemize{

  \item{operation_id}{ a unique id for the edit}

  \item{timestamp}{ in POSIXct format, to the nearest ms}

  \item{user_id}{ numeric id for the user responsible for the edit}

  \item{is_merge}{ whether it was a merge or a split}

  \item{user_name}{ as a string}

  }

  In addition when \code{filtered=FALSE}, \code{in_neuron} \code{is_relevant}

  In addition when \code{root_ids=TRUE}, \code{before_root_ids}
  \code{after_root_ids}, as space separated strings.
}
\description{
Summarise edits for one or more FlyWire segment ids
}
\details{
When \code{filtered=TRUE} the following are removed \itemize{

  \item edits that are outside of the given root id - meaning these edits are
  in the history of the cell but the affected part of the arbour was since
  cut off.

  \item edits that were since undone by an inverse operation (often temporary
  splits made while reviewing a neuron.)

  }
}
\examples{
\donttest{
flywire_change_log("720575940619010932")
flywire_change_log("720575940619010932", root_ids = TRUE)
flywire_change_log("720575940619010932", filtered = FALSE)
# with a flywire URL
u="https://ngl.flywire.ai/?json_url=https://globalv1.flywire-daf.com/nglstate/5409525645443072"
flywire_change_log(u)
}

}
