% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-arrow.R
\name{flywire_connectome_data}
\alias{flywire_connectome_data}
\alias{flywire_connectome_data_version}
\title{Access precomputed flywire connectivity data}
\usage{
flywire_connectome_data(
  type = c("syn", "pre", "post"),
  version = NULL,
  cached = TRUE,
  ...
)

flywire_connectome_data_version(set = NULL, default = NA)
}
\arguments{
\item{type}{Character vector specifying the kind of data}

\item{version}{Optional CAVE version. The default value of \code{NULL} uses
the latest data dump available unless
\code{options(fafbseg.flywire_connectome_data_version)} has been set (which
you can conveniently do using
\code{\link{flywire_connectome_data_version}()}).}

\item{cached}{When version is \code{NULL} whether to use a cached value
(lasting 1 hour) of the latest available version.}

\item{...}{Additional arguments passed to \code{arrow::open_dataset}.}

\item{set}{When \code{set=<number>} is passed as an argument the specified
data version will be used going forwards in this session as the default.
This is achieved by setting the
\code{fafbseg.flywire_connectome_data_version} option. When \code{set=NA}
is specified then the option is cleared. When \code{set=FALSE}, the latest
version on disk will be returned regardless of the value of
\code{options("fafbseg.flywire_connectome_data_version"))}. See examples.}

\item{default}{A version to return when no other information is available.
Defaults to \code{NA} to indicate no version information available.}
}
\value{
An arrow object that you can use with \code{dplyr} verbs like
  \code{filter} in order to find neurons/connectivity data of interest.

An integer version number \emph{or} a list with the previous value of
  \code{options(fafbseg.flywire_connectome_data_version)} when
  \code{set=<number>}.
}
\description{
\code{flywire_connectome_data} returns a \code{\link{dplyr}}
  compatible connection to connectivity dumps on disk.

\code{flywire_connectome_data_version} sets the integer version
  number of the preferred flywire connectome data dump or returns the
  currently version.
}
\details{
This depends on precomputed data dumps prepared periodically by Sven
  Dorkenwald. You can download the public release version using the function
  \code{\link{download_flywire_release_data}}.

  You can download other versions from Sven's Google drive folder. See
  \href{https://flywire-forum.slack.com/archives/C01M4LP2Y2D/p1644529750249139}{this FlyWire
   Slack message} for more details and the URL.

Two pieces of information are used to determine the \emph{version} when it
  is queried. First the value of
  \code{options(fafbseg.flywire_connectome_data_version)}, second the latest
  available version of the connectivity dumps provided by
  \code{flywire_connectome_data()}.
}
\examples{
\donttest{
# latest available version/
syn=try(flywire_connectome_data('syn'), silent=TRUE)
syn450=try(flywire_connectome_data('syn', version=450), silent=TRUE)
if(!inherits(syn450, 'try-error')) {
syn450
syn450$metadata

dl4ds <- syn450 \%>\%
  filter(pre_pt_root_id==flywire_ids("DL4_adPN_L", version=450, integer64 = TRUE)) \%>\%
  collect()
}

}
\dontrun{
# report active connectome dump version (defaults to most recent available)
flywire_connectome_data_version()

# use the June 2023 public release version as the default
flywire_connectome_data_version(set=630)
# confirm this is the default
flywire_connectome_data_version()

# check the latest version on disk
flywire_connectome_data_version(set=FALSE)

# stop defaulting to specific version (therefore using the latest on disk)
flywire_connectome_data_version(set=NA)
flywire_connectome_data_version()
}
}
\seealso{
\code{\link{download_flywire_release_data}}
}
