% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosyn.R
\name{flywire_dcvs}
\alias{flywire_dcvs}
\title{Get predicted dense core vesicle locations in the flywire dataset}
\usage{
flywire_dcvs(
  rootid,
  url = "https://radagast.hms.harvard.edu/fafb_test/dcv/for_segment",
  simplify = TRUE,
  return = c("parsed", "text", "response"),
  token = NULL,
  simplifyVector = TRUE,
  include_headers = FALSE,
  OmitFailures = TRUE,
  ...
)
}
\arguments{
\item{rootid}{flywire rootid/rootids}

\item{url}{Full URL for brainmaps API endpoint}

\item{simplify}{logical, if \code{FALSE} each rootid is a separate set of two data frames (one for DCV positions, one for the neuron's synapses).
Else, a list of two combined data frames is returned.}

\item{return}{One of "parsed", "text" (for raw JSON), or "response"}

\item{token}{Optional chunkedgraph token (otherwise the default one for the
current segmentation will be used).}

\item{simplifyVector}{Whether to use \code{jsonlite::simplifyVector}}

\item{include_headers}{Whether to include basic headers from the http request
as attributes on the parsed JSON object (default \code{TRUE}) when 
\code{parse.json=TRUE}.}

\item{OmitFailures}{logical, if \code{TRUE} then requests that result in a 500 error are dropped and a warning is displayed
but not an error.}

\item{...}{additional arguments passed to the \code{httr::{RETRY}} function.
This may include a \code{\link[httr]{config}} list other named parameters
etc.}
}
\value{
A list, where the first entry contains DCV locations and the second the synapses for the given rootid, with the nearest DCV precalculated for each synapse..
}
\description{
Preliminary dense core vesicle (DCV) detection results from the Lee group, Stephan Gerhard and Minsu Kim.
}
\examples{
\donttest{
data = flywire_dcvs("720575940629166904")
dcv = data$dcv
synapse = data$syns
}
}
