% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2cache.R
\name{flywire_has_l2cache}
\alias{flywire_has_l2cache}
\title{Check whether an L2 cache is available}
\usage{
flywire_has_l2cache(
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production")
)
}
\arguments{
\item{datastack_name}{defaults to the value selected by
\code{\link{choose_segmentation}} and to "flywire_fafb_production" when
that is missing. See \url{https://global.daf-apis.com/info/} for other
options.}
}
\value{
Logical scalar indicating whether the current datastack has an
  associated level 2 cache.
}
\description{
Check whether an L2 cache is available
}
