% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2cache.R
\name{flywire_l2cache_metadata}
\alias{flywire_l2cache_metadata}
\title{Return FlyWire level 2 cache metadata}
\usage{
flywire_l2cache_metadata(
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production")
)
}
\arguments{
\item{datastack_name}{defaults to the value selected by
\code{\link{choose_segmentation}} and to "flywire_fafb_production" when
that is missing. See \url{https://global.daf-apis.com/info/} for other
options.}
}
\value{
A named list mapping available level 2 cache attributes to their
  declared data types.
}
\description{
Return FlyWire level 2 cache metadata
}
