% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2cache.R
\name{flywire_l2volume}
\alias{flywire_l2volume}
\title{Sum level 2 volumes for one or more FlyWire root ids}
\usage{
flywire_l2volume(
  rootids,
  cache = TRUE,
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production"),
  ...
)
}
\arguments{
\item{rootids}{One or more FlyWire root ids.}

\item{cache}{Whether to cache rootid-level volume totals on disk.}

\item{datastack_name}{defaults to the value selected by
\code{\link{choose_segmentation}} and to "flywire_fafb_production" when
that is missing. See \url{https://global.daf-apis.com/info/} for other
options.}

\item{...}{Additional arguments passed to \code{\link{flywire_l2attributes}}
and to recursive calls when \code{rootids} has length greater than one.}
}
\value{
A numeric scalar volume in nm^3 for one input root id, or a named
  vector when \code{rootids} has length greater than one.
}
\description{
Sum level 2 volumes for one or more FlyWire root ids
}
