% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosyn.R
\name{flywire_ntpred}
\alias{flywire_ntpred}
\alias{print.ntprediction}
\title{Return raw neurotransmitter prediction results for output of flywire neuron}
\usage{
flywire_ntpred(
  x,
  cleft.threshold = 0,
  remove_autapses = TRUE,
  local = NULL,
  cloudvolume.url = NULL
)

\method{print}{ntprediction}(x, ...)
}
\arguments{
\item{x}{A single root id as a string OR a \code{data.frame} of output
(downstream) partners returned by \code{flywire_partners}.}

\item{cleft.threshold}{A threshold for the cleft score calculated by Buhmann
et al 2019 (default 0, we have used 30-100 to increase specificity)}

\item{remove_autapses}{For \code{flywire_partner_summary} whether to remove
autapses (defaults to TRUE)}

\item{local}{path to SQLite synapse data. Evaluated by
\code{fafbseg:::local_or_google}. Work in progress. Default is to download
this data and place it in \code{~/projects/JanFunke}.}

\item{cloudvolume.url}{The segmentation source URL for cloudvolume. Normally
you can ignore this and rely on the default segmentation chosen by
\code{\link{choose_segmentation}}}

\item{...}{additional arguments passed to \code{\link{print}}}
}
\value{
A \code{data.frame} of neurotransmitter predictions
}
\description{
the \code{print.ntprediction} method provides a quick summary of
  the neurotransmitter prediction for all output synapses.
}
\examples{
\donttest{
# an olfactory projection neuron
flywire_ntpred("720575940615237849")
# alternatively
\dontrun{
flywire_ntpred(flywire_xyz2id(cbind(116923, 61378, 1474), rawcoords = T))
}
}
}
\seealso{
Other automatic-synapses:
\code{\link[=flywire_adjacency_matrix]{flywire_adjacency_matrix()}},
\code{\link[=flywire_neurons_add_synapses]{flywire_neurons_add_synapses()}},
\code{\link[=flywire_ntplot]{flywire_ntplot()}},
\code{\link[=flywire_partners]{flywire_partners()}}
}
\concept{automatic-synapses}
