% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-nuclei.R
\name{flywire_nuclei}
\alias{flywire_nuclei}
\title{Queries for information about flywire nuclei (helpful for finding somata)}
\usage{
flywire_nuclei(rootids = NULL, nucleus_ids = NULL, ...)
}
\arguments{
\item{rootids}{Character vector specifying one or more flywire rootids. As a
convenience for \code{flywire_partner_summary} this argument is passed to
\code{\link{ngl_segments}} allowing you to pass in flywire URLs.}

\item{...}{Additional arguments passed to \code{\link{flywire_cave_query}}}
}
\value{
A data.frame containing information about nuclei including \itemize{

  \item id nucleus id

  \item pt_position the XYZ position of the centre of the nucleus (in nm)

  \item pt_supervoxel_id corresponding to the nucleus may be missing if the
  segmentation is disrupted at the location of the nucleus e.g. because of
  masking issues, missing sections etc.

  \item pt_root_id The current root id (when the \code{pt_position} maps onto
  the segmentation.)

  \item volume the volume in cubic microns of the nucleus

  }
}
\description{
Queries for information about flywire nuclei (helpful for finding somata)
}
\examples{
\donttest{
# an example where there are two nucleus matches
flywire_nuclei(flywire_xyz2id(c(120152, 22864, 3564), rawcoords = TRUE))
}
}
