% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-arrow.R
\name{flywire_partner_summary2}
\alias{flywire_partner_summary2}
\title{Rapid flywire connectivity summaries using cached connectome data}
\usage{
flywire_partner_summary2(
  ids,
  partners = c("outputs", "inputs"),
  add_cell_types = TRUE,
  by.roi = FALSE,
  summarise = FALSE,
  threshold = 0,
  version = NULL
)
}
\arguments{
\item{ids}{Root ids to query (passed to \code{\link{flywire_ids}})}

\item{partners}{Whether to fetch input or output synapses or both.}

\item{add_cell_types}{Whether to add cell type information to the result}

\item{by.roi}{Whether to break the connectivity down into rows for each
neuropil region containing synapses.}

\item{summarise}{Whether to collapse down the results for multiple query
neurons into a single entry for each partner neuron.}

\item{threshold}{For \code{flywire_partner_summary} only return partners with
greater than this number of connections to the query neuron(s) (default of
0 returns all connections)}

\item{version}{Optional CAVE version. The default value of \code{NULL} uses
the latest data dump available unless
\code{options(fafbseg.flywire_connectome_data_version)} has been set (which
you can conveniently do using
\code{\link{flywire_connectome_data_version}()}). The special version of
\code{"783.2"} will use the 2025 (Princeton) synapse data release.}
}
\value{
A data.frame
}
\description{
Rapid flywire connectivity summaries using cached connectome data
}
\details{
Note that the threshold is applied to each row left after any
  grouping operations. Therefore when \code{by.roi=TRUE} only neuropil
  regions exceeding this threshold will be returned.

  CAVE specifies versions (effectively timestamps) for the connectome data.
  Every so often Sven makes a dump of the connectivity and synapse
  information for all proofread neurons. At this point, the only versions in
  use are 783 (released with the published articles in Nature) and 630
  (released with the preprint).

  The 783 release is available with both the original connectivity based on
  Buhmann et at 2021 (still the default) as well as new v2 synapses released
  in 2025 (version='783.2')
}
\examples{
\dontrun{
flywire_partner_summary2('DA2_lPN', partners='out')
flywire_partner_summary2('DA2_lPN', partners='out', summarise=T)
flywire_partner_summary2('DA2_lPN', partners='out', summarise=T, by.roi=T)

flywire_partner_summary2('DA2_lPN', partners='out', summarise=T,
  by.roi=T, add_cell_types=F) \%>\%
  filter(!grepl("AL", neuropil)) \%>\%
  group_by(post_pt_root_id) \%>\%
  summarise(weight = sum(weight), top_np = neuropil[1]) \%>\%
  arrange(desc(weight)) \%>\%
  # nb version = TRUE will use ensure that ids match the default CAVE version
  add_celltype_info(version=TRUE)
}
}
