% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudvolume.R
\name{chunkedgraph_token}
\alias{chunkedgraph_token}
\alias{flywire_set_token}
\title{Record (and if necessary create) a FlyWire chunkedgraph token}
\usage{
chunkedgraph_token(url = NULL, cached = TRUE)

flywire_set_token(token = NULL, domain = NULL)
}
\arguments{
\item{url}{A URL or domain name which defines the scope of the token}

\item{cached}{Whether to use a cached version of the token (default yes for
speed, but set to FALSE to reload after writing a new token to disk).}

\item{token}{Optional character vector containing your token. If missing, a
new token will be requested (note that this will invalidate your previous
token)}

\item{domain}{Optional fully qualified domain name, which will be prepended
to the filename in which the token will be stored.}
}
\value{
character vector containing the token (typically 32-44 bytes)

The path to the file storing the token (invisibly)
}
\description{
\code{chunkedgraph_token} returns the current chunked graph
  token, optionally for a specified URL / domain name.

Writes a token to a standard location on disk so that it will be found by the
cloudvolume python package as well as the fafbseg package and used to
authenticate to \url{https://flywire.ai}.
}
\details{
Since cloudvolume 3.10.0 April 2021, the recommended token filenames look like
\itemize{

\item \code{'cave-secret.json'}

\item \code{'cave-secret.json'}

}
}
\examples{
\dontrun{
# Will open browser to get new token
flywire_set_token()
# Writes a known token to correct location
flywire_set_token("2f88e16c4f21bfcb290b2a8288c05bd0")
}
}
