% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release-data.R
\name{flywire_sirepo_file}
\alias{flywire_sirepo_file}
\alias{flywire_sirepo_file_memo}
\title{Read or return path to FlyWire annotations manuscript supplementary file}
\usage{
flywire_sirepo_file(
  p,
  mustWork = NA,
  read = FALSE,
  version = c(783L, 630L),
  ...
)

flywire_sirepo_file_memo(
  p,
  mustWork = NA,
  read = FALSE,
  version = c(783L, 630L),
  ...
)
}
\arguments{
\item{p}{Relative path to file within flywire_annotations repository}

\item{mustWork}{Whether the path must exists (default \code{NA} =>
\code{TRUE} when reading the file)}

\item{read}{Whether to read the file. Either a logical value or a function.
When \code{TRUE} and \code{p} is a tsv or csv file a default read function
is used (see details).}

\item{version}{An integer CAVE materialisation version (see
\code{\link{flywire_connectome_data_version}})}

\item{...}{Additional arguments passed to the function determined by the
\code{read} argument (typically
\code{data.table::\link[data.table]{fread}}).}
}
\value{
A path or (when \code{read=TRUE} or a function) the result of reading
  the file (a \code{data.table} for csv/tsv files).
}
\description{
\code{flywire_sirepo_file_memo()} is a memoised version with a 5
minute timeout
}
\details{
When \code{TRUE} and \code{p} is a tsv or csv file them the
  \code{data.table::\link{fread}} function is used in order to ensure that 64
  bit integers are correctly parsed. The default behaviour is to read ids as
  character vectors but this can be overridden (see examples).
}
\examples{
\dontrun{
annpath=flywire_sirepo_file('supplemental_files/Supplemental_file1_annotations.tsv',
  read=FALSE)
# read in annotation file
anns=flywire_sirepo_file('supplemental_files/Supplemental_file1_annotations.tsv',
  read = TRUE)
# read in annotation file with ids as 64 bit integers rather than strings
anns=flywire_sirepo_file('supplemental_files/Supplemental_file1_annotations.tsv',
  read = TRUE, integer64="integer64")
# same but memoised to avoid checking github / re-reading file
anns=flywire_sirepo_file_memo('supplemental_files/Supplemental_file1_annotations.tsv',
  read = TRUE, integer64="integer64")
}
}
