% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_updateids}
\alias{flywire_updateids}
\title{Update flywire root ids to any timestamp using XYZ position or supervoxel ids}
\usage{
flywire_updateids(
  x,
  svids = NULL,
  xyz = NULL,
  rawcoords = FALSE,
  voxdims = c(4, 4, 40),
  cache = NA,
  version = NULL,
  timestamp = NULL,
  Verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Current root ids}

\item{svids}{optional supervoxel ids}

\item{xyz}{optional xyz locations in any form understood by
\code{\link{xyzmatrix}}}

\item{rawcoords}{whether the input values are raw voxel indices or in nm}

\item{voxdims}{voxel dimensions in nm used to convert raw coordinates. The
default value uses the \code{\link{flywire_voxdims}} function to identify
the value for the current segmentation (usually with success).}

\item{cache}{Whether to cache supervoxel id to root id mappings. The default
value of \code{NA} will do this when a version or timestamp argument is
specified. See \code{\link{flywire_rootid}} for details.}

\item{version}{An optional CAVE materialisation version number. See details
and examples.}

\item{timestamp}{An optional timestamp as a string or POSIXct, interpreted as
UTC when no timezone is specified.}

\item{Verbose}{Whether to print a message to the console when updates are
required.}

\item{...}{Additional arguments passed to \code{\link{flywire_islatest}} or
\code{\link{flywire_latestid}}}
}
\value{
A vector of the same form as \code{x} with updated ids.
}
\description{
\code{flywire_updateids} updates root ids to the latest version
  (or any arbitrary integer materialisation version / timestamp that you
  specify). To be as efficient as possible, it will use supervoxel ids, XYZ
  positions or failing that the slower \code{\link{flywire_latestid}}.

  As of November 2022 the default behaviour will use a per-session supervoxel
  id to root id cache in order to speed up repeated lookups for the same root
  id / supervoxel id pairs at a given timestamp/version. See
  \code{\link{flywire_rootid}} for further details.
}
\examples{

kcs=data.frame(
rootid=c("720575940602553568", "720575940602564320", "720575940602605536"),
xyz=c("(159284,42762,3594)", "(159035,41959,3594)", "(157715,44345,3594)")
)
# update root ids
kcs$rootid=flywire_updateids(kcs$rootid, xyz=kcs$xyz)
}
\seealso{
Other flywire-ids:
\code{\link[=flywire_islatest]{flywire_islatest()}},
\code{\link[=flywire_last_modified]{flywire_last_modified()}},
\code{\link[=flywire_latestid]{flywire_latestid()}},
\code{\link[=flywire_leaves]{flywire_leaves()}},
\code{\link[=flywire_rootid]{flywire_rootid()}},
\code{\link[=flywire_xyz2id]{flywire_xyz2id()}}
}
\concept{flywire-ids}
