% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge_info.R
\name{make_merge_graph}
\alias{make_merge_graph}
\alias{merge_graph_components}
\title{Graph representation of the agglomeration (merge) between raw segments}
\usage{
make_merge_graph(x, n = NA)

merge_graph_components(g, n = NA)
}
\arguments{
\item{x}{The \code{mergeinfo} \code{data.frame} generated by
\code{\link{read_mergeinfo}}}

\item{n}{For testing purposes, the number of rows of merge information to
process - the default processes all rows.}

\item{g}{Either a graph generated by \code{make_merge_graph} or the
\code{mergeinfo} \code{data.frame} generated by
\code{\link{read_mergeinfo}}.}
}
\value{
\code{make_merge_graph} returns an \code{igraph} object,
  \code{merge_graph_components} returns a list with elements describing the
  mapping between segment ids and merge groups.
}
\description{
\code{make_merge_graph} generates a graph representation of all
  agglomeration merges.

\code{merge_graph_components} uses the full merge graph to find
  each connected subgraph corresponding to isolated merge groups.
}
\examples{
\dontrun{
g=make_merge_graph(mergeinfo)
gc=merge_graph_components(g)
str(gc)
}
}
