% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshparty-reticulate.R
\name{meshparty_skeletonize}
\alias{meshparty_skeletonize}
\title{Skeletonize neurons using meshparty python library}
\usage{
meshparty_skeletonize(segments, savedir = NULL, invalidation_d = 12000, ...)
}
\arguments{
\item{segments}{neuron ids in any form understood by
\code{\link{ngl_segments}} OR paths to obj files already saved to disk.}

\item{savedir}{Where to save SWC files (defaults to temporary directory)}

\item{invalidation_d}{Distance parameter (nm) controlling skeletonisation
level of detail. See meshparty docs.}

\item{...}{additional arguments passed to \code{cloudvolume_save_obj}}
}
\value{
A character vector containing the path to one or more SWC files. Note
  that these SWCs will be calibrated in µm even though the input data are in
  nm.
}
\description{
Skeletonize neurons using meshparty python library
}
\examples{
\dontrun{
meshparty_skeletonize("720575940614134045")
meshparty_skeletonize("720575940614134045.obj")
}
}
