% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_add_colours}
\alias{ngl_add_colours}
\title{Add colours to the neuroglancer scene}
\usage{
ngl_add_colours(x, colours, layer = NULL)
}
\arguments{
\item{x}{neuroglancer scene}

\item{colours}{A dataframe with two columns, where the first is the id and
the second is the colour, OR a character vector of colours named by the ids
or one colour which would be added to all the displayed neurons. Will
choose a layer of type segmentation_with_graph if one exists.}
}
\value{
A neuroglancer scene object (see \code{\link{ngl_decode_scene}})
}
\description{
Add colours to the neuroglancer scene
}
\details{

}
\examples{
fw_url=with_segmentation('flywire', getOption('fafbseg.sampleurl'))
ngl_add_colours(fw_url, colours=c("1"="red"))
}
