% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng-annotations.R
\name{ngl_annotations}
\alias{ngl_annotations}
\title{Extract annotations from a neuroglancer scene into a dataframe}
\usage{
ngl_annotations(
  x,
  layer = NULL,
  types = c("point", "line"),
  points = c("collapse", "expand", "list")
)
}
\arguments{
\item{x}{A neuroglancer scene or URL (passed to
\code{\link{ngl_decode_scene}} as necessary)}

\item{layer}{Optional index vector specifying the layers within a scene from
which to extract annotations. It is probably safest to use a character
vector of layer names (what appears in neuroglancer). When missing all
annotation layers are processed.}

\item{types}{Which annotation types to process (currently only points and
lines by default)}

\item{points}{What to do with point coordinates.}
}
\value{
A data.frame with columns defined by the contents of the annotation
  layer and the \code{types}/\code{points} arguments. Additional attributes are stored
}
\description{
Extract annotations from a neuroglancer scene into a dataframe
}
\examples{
\donttest{
u="https://ngl.flywire.ai/?json_url=https://globalv1.flywire-daf.com/nglstate/5963849085747200"
adf=ngl_annotations(u)
str(attr(adf, 'ann_attrs'))
}
}
