% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_blank_scene}
\alias{ngl_blank_scene}
\title{Return a blank neuroglancer scene based on a specified segmentation}
\usage{
ngl_blank_scene(release = NULL, return.url = FALSE)
}
\arguments{
\item{release}{character vector specifying a released segmentation via a
known short name or a sample neuroglancer URL.}

\item{return.url}{Whether to return a URL rather than a \code{ngscene}
object.}
}
\description{
defaults to the current segmentation defined by
  \code{\link{choose_segmentation}} when \code{release=NULL}
}
\examples{
# blank scene for current segmentation
ngl_blank_scene()
# add a specific id
ngl_blank_scene()+"720575940623755722"
# a different segmentation
ngl_blank_scene("202004")
u=ngl_blank_scene("202004", return.url=TRUE)
\dontrun{
u=ngl_blank_scene("sandbox")
browseURL(u)
}
}
\seealso{
\code{\link{choose_segmentation}}

Other neuroglancer-urls:
\code{\link[=flywire_ids]{flywire_ids()}},
\code{\link[=flywire_scene]{flywire_scene()}},
\code{\link[=ngl_decode_scene]{ngl_decode_scene()}},
\code{\link[=ngl_encode_url]{ngl_encode_url()}},
\code{\link[=ngl_segments]{ngl_segments()}},
\code{\link[=open_fafb_ngl]{open_fafb_ngl()}}
}
\concept{neuroglancer-urls}
