% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_decode_scene}
\alias{ngl_decode_scene}
\alias{ngscene}
\title{Decode and manipulate Neuroglancer scenes (from URLs or JSON blocks)}
\usage{
ngl_decode_scene(
  x,
  return.json = FALSE,
  simplifyVector = TRUE,
  simplifyDataFrame = FALSE,
  ...
)
}
\arguments{
\item{x}{Character vector containing single Neuroglancer URL or a json block}

\item{return.json}{When \code{TRUE} extracts the JSON block in a URL does not
parse it to an R list}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{simplifyDataFrame}{coerce JSON arrays containing only records (JSON objects) into a data frame}

\item{...}{additional arguments passed to \code{\link[jsonlite]{fromJSON}}}
}
\value{
An R list with additional class \code{ngscene} describing the scene,
  or, when \code{return.json=TRUE}, a character vector.
}
\description{
\code{ngl_decode_scene} takes a Neuroglancer scene from your web
  browser and turns it into an R \code{list} object that can be
  programmatically manipulated e.g. to add/remove segments. Manipulate these
  scenes with \code{\link{ngl_segments}}, \code{\link{ngl_layers}}. See
  \code{\link{ngl_encode_url}} to turn a scene back into a URL to open in your
  browser.
}
\examples{

\donttest{
# get sample FlyWire URL
fw_url=with_segmentation('flywire31', getOption('fafbseg.sampleurl'))
# only a 0 (dummy) segment id present
ngl_segments(fw_url)
#
fw_sc=ngl_decode_scene(fw_url)
fw_sc
# add two segments
fw_sc=fw_sc+c("720575940621039145", "720575940626877799")
ngl_segments(fw_sc)
# remove that 0 segment
fw_sc=fw_sc-0
ngl_segments(fw_sc)
# repeated segments are ignored i.e. no duplicates
ngl_segments(fw_sc+"720575940621039145")
# convert back to a URL, nb this depends on choose_segmentation
ngl_encode_url(fw_sc)

\dontrun{
# open in your default browser
browseURL(ngl_encode_url(fw_sc))
}
}

\dontrun{
ngl_decode_scene("<someneuroglancerurl>")

# decode scene from URL currently on clipboard
scene=ngl_decode_scene(clipr::read_clip())

# open a Neuroglancer URL in CATMAID
ngu="<someurl>"
library(elmr)
open_fafb(ngl_decode_scene(ngu))
# Or store the URL rather than opening it
cmu=open_fafb(ngl_decode_scene(ngu), open=FALSE)
}
}
\seealso{
\code{\link[utils]{URLdecode}}, \code{\link[jsonlite]{fromJSON}}

Other neuroglancer-urls:
\code{\link[=flywire_ids]{flywire_ids()}},
\code{\link[=flywire_scene]{flywire_scene()}},
\code{\link[=ngl_blank_scene]{ngl_blank_scene()}},
\code{\link[=ngl_encode_url]{ngl_encode_url()}},
\code{\link[=ngl_segments]{ngl_segments()}},
\code{\link[=open_fafb_ngl]{open_fafb_ngl()}}
}
\concept{neuroglancer-urls}
