% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_encode_url}
\alias{ngl_encode_url}
\alias{as.character.ngscene}
\title{Encode scene information into a neuroglancer URL}
\usage{
ngl_encode_url(body, baseurl = NULL, auto_unbox = TRUE, ...)

\method{as.character}{ngscene}(x, ...)
}
\arguments{
\item{body}{A text file or character vector with JSON data or an R list
object of class \code{ngscene}.}

\item{baseurl}{A URL specifying the neuroglancer server (if missing, uses the
URL from which \code{body} was decoded if that was recorded or, failing
that, \code{options("fafbseg.sampleurl")}). You can use any neuroglancer
URL as will be appropriately truncated if it encodes scene information.}

\item{auto_unbox}{For expert use only. See \code{\link[jsonlite]{toJSON}} for
details.}

\item{...}{Additional arguments for \code{\link[jsonlite]{toJSON}}}

\item{x}{the \code{ngscene} object to be converted to a URL}
}
\value{
Character vector containing encoded URL
}
\description{
\code{ngl_encode_url} converts an R list containing a
  neuroglancer scene into a URL that you can open in your browser.

\code{as.character.ngscene} is another way to convert a
  neuroglancer scene object to a URL.
}
\details{
We take pains to ensure that entries that neuroglancer expects to be
  JSON arrays are (including \code{segments} and \code{hiddenSegments}) are
  always mapped to a JSON array (even when length 1).

  The default baseurl depends on the current segmentation chosen by
  \code{\link{choose_segmentation}}.
}
\examples{
\donttest{
# get sample FlyWire URL
fw_url=with_segmentation('flywire31', getOption('fafbseg.sampleurl'))
# only a 0 (dummy) segment id present
ngl_segments(fw_url)
#
fw_sc=ngl_decode_scene(fw_url)
# add a segment
fw_sc$layers[[2]]$segments=union(fw_sc$layers[[2]]$segments,
  "720575940626877799")
# convert back to a URL, nb this depends on choose_segmentation
ngl_encode_url(fw_sc)
# another way to do this, which long time R users may find more intuitive
as.character(fw_sc)

\dontrun{
# open in your default browser
browseURL(ngl_encode_url(fw_sc))
# ... or
browseURL(as.character(fw_sc))
}
}

\dontrun{
# copy JSON scene information from {} symbol at top right of neuroglancer
# now make a permanent URL for the scene
ngl_encode_url(clipr::read_clip())
}
}
\seealso{
\code{\link{URLencode}}, \code{\link{open_fafb_ngl}},
  \code{\link[jsonlite]{toJSON}}

Other neuroglancer-urls:
\code{\link[=flywire_ids]{flywire_ids()}},
\code{\link[=flywire_scene]{flywire_scene()}},
\code{\link[=ngl_blank_scene]{ngl_blank_scene()}},
\code{\link[=ngl_decode_scene]{ngl_decode_scene()}},
\code{\link[=ngl_segments]{ngl_segments()}},
\code{\link[=open_fafb_ngl]{open_fafb_ngl()}}
}
\concept{neuroglancer-urls}
