% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ngl_layers}
\alias{ngl_layers}
\title{Extract and manipulate layers in a neuroglancer scene}
\usage{
ngl_layers(x, subset = NULL)
}
\arguments{
\item{x}{a neuroglancer scene object (see \code{\link{ngscene}})}

\item{subset}{an expression (evaluated in the style of subset.dataframe) which defined}
}
\description{
Extract and manipulate layers in a neuroglancer scene
}
\examples{
\donttest{
u="https://ngl.flywire.ai/?json_url=https://globalv1.flywire-daf.com/nglstate/5409525645443072"
sc=ngl_decode_scene(u)
sc
names(ngl_layers(sc))
str(ngl_layers(sc))
}
}
