% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{read_brainmaps_meshes}
\alias{read_brainmaps_meshes}
\title{Read 3D meshes via the brainmaps API}
\usage{
read_brainmaps_meshes(x, volume = getOption("fafbseg.brainmaps.volume"),
  meshName = getOption("fafbseg.brainmaps.meshName"), ...)
}
\arguments{
\item{x}{Vector of integer segment ids}

\item{volume}{String identifier for the volume containing segments}

\item{meshName}{String identifier for the meshes}

\item{...}{Additional arguments passed to \code{\link{brainmaps_fetch}}}
}
\value{
A \code{\link[rgl]{mesh3d}} object
}
\description{
Read 3D meshes via the brainmaps API
}
\examples{
\dontrun{
segs=find_merged_segments(7186840767)
samplemesh=read_brainmaps_meshes(segs)
sampleskel=read_segments2(segs)
dot3d(samplemesh, col='grey')
plot3d(sampleskel, lwd=2)

# or compare mesh and skeleton colouring points by distance from skeleton
compare_ng_neuron(samplemesh, sampleskel, pointsize=1, sample_dots = 0.3)
}
}
\seealso{
\code{\link{read_segments2}} to read skeleton fragments and
  \code{\link{brainmaps_listfragments}} (to identify the fragments that must
  be read). See \code{\link{compare_ng_neuron}} to compare skeletons and 3D
  meshes.
}
