% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge_info.R
\name{read_mergeinfo}
\alias{read_mergeinfo}
\title{Read CSV files containing raw agglomeration (merge) information}
\usage{
read_mergeinfo(x, pattern = "10\\\\.csv$", voxdims = c(16, 16, 40), ...)
}
\arguments{
\item{x}{Input directory}

\item{pattern}{Optional pattern to identify CSV files}

\item{voxdims}{Optional voxel dimensions to convert raw pixel coordinates
into physical (nm) coordinates}

\item{...}{Additional arguments passed to \code{readr::\link{read_csv}}}
}
\value{
a \code{data.frame} with columns \itemize{

  \item{id1,id2}{ Segment ids to be merged}

  \item{x,y,z}{ Location (in nm) of merge point}

  }
}
\description{
Read CSV files containing raw agglomeration (merge) information
}
\details{
Peter Li provided 10 CSV files with merge information in Aug 2018
  against which this function has been tested. This merge information has
  been processed to identify which groups segments should be merged together
  in the \code{\link[fafbsegdata]{fafbsegdata}} package, which can be
  accessed using the \code{\link{find_merged_segments}} function.
}
\examples{
\dontrun{
csvdir="~/projects/fafbseg/fafb14_v00c_split3xfill2x_merges/ffnreseg16nm_ms1000_md0.02_c0.6_iou0.7/"
mergeinfo <- read_mergeinfo(csvdir)

# Show 3D position of merge locations for a large merge group containing a
# tangential cell (widefield visual interneuron)
tang_segs <- find_merged_segments(22139217567)
tang_mi <- subset(mergeinfo, id1\%in\% tang_segs | id2 \%in\%tang_segs)
points3d(xyzmatrix(tang_mi))

# plot brain surface for context
library(elmr)
plot3d(FAFB)
}
}
\seealso{
\code{\link{find_merged_segments}},
  \code{\link[fafbsegdata]{fafbsegdata}} package
}
