% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-all.R
\name{read_ng_dump}
\alias{read_ng_dump}
\title{Read neuroglancer data dump from disk / web}
\usage{
read_ng_dump(x, ...)
}
\arguments{
\item{x}{Path to directory, zip file, or set of data files}

\item{...}{Additional argument passed to \code{\link{read_ng_raw}}}
}
\value{
A list of class \code{ng_raw_list} with additional metadata as
  attributes.
}
\description{
Read neuroglancer data dump from disk / web
}
\details{
A neuroglancer data dump consists of a mix of JSON and custom binary
  data files, which together define mesh data for (fragments) of neurons
  together with associated metadata. See \code{\link{fetch_all_curl}} for how
  to prepare such a data dump.
}
\examples{
\dontrun{
fetch_all_curl("all_curl.sh", outdir="alldata",
  regex="brainmaps.googleapis.com", fixed=TRUE)
meshdata=read_ng_dump("alldata")
m=as.mesh3d(meshdata)
shade3d(m, col='red')

# can also be a (remote) zip file
meshdata=read_ng_dump("https://myfiles.com/myneuron.zip")
}
}
\seealso{
\code{\link{fetch_all_curl}}, \code{\link{read_ng_raw}}
}
