% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng-raw-io.R
\name{read_ng_raw}
\alias{read_ng_raw}
\title{Read raw neuroglancer data - currently only supports mesh format}
\usage{
read_ng_raw(x, read_data = TRUE, Verbose = FALSE)
}
\arguments{
\item{x}{Path to one or more files}

\item{read_data}{Whether to read the data (default when \code{TRUE}) or just
the header}

\item{Verbose}{Whether to print some status messages (default \code{FALSE})}
}
\value{
An object of class 'ng_raw_list' containing one or more chunks of
  data of class 'ng_raw'. When \code{x} contains multiple files, all the
  chunks are merged into a single list.
}
\description{
Read raw neuroglancer data - currently only supports mesh format
}
\examples{
\dontrun{
res <- read_ng_raw("meshdata/chunk00789.raw")
resh <- read_ng_raw("meshdata/chunk00789.raw", read_data=FALSE)
resl <- read_ng_raw(dir("meshdata", full.names = TRUE)
}
}
\references{
See \url{https://github.com/google/neuroglancer}
}
