% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudvolume.R
\name{set_chunkedgraph_token}
\alias{set_chunkedgraph_token}
\title{Record (and if necessary create) a FlyWire chunkedgraph token}
\usage{
set_chunkedgraph_token(token = NULL)
}
\arguments{
\item{token}{Optional character vector containing your token. If missing, a
new token will be requested}
}
\value{
The path to the file storing the token (invisibly)
}
\description{
Writes a token to a standard location on disk so that it will be found by the
cloudvolume python package as well as the fafbseg package and used to
authenticate to \url{https://flywire.ai}.
}
\examples{
\dontrun{
# Will open browser to get new token
set_chunkedgraph_token()
# Writes a known token to correct location
set_chunkedgraph_token("2f88e16c4f21bfcb290b2a8288c05bd0")
}
}
