% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tabify_coords}
\alias{tabify_coords}
\title{Convert coordinates to tab separated values. Useful for copy/paste to
Seatable}
\usage{
tabify_coords(xyz = NULL, FUN = NULL, write_clip = NULL)
}
\arguments{
\item{xyz}{3D Coordinates in any form compatible with
\code{\link{xyzmatrix}}. When missing these are read from the clipboard.}

\item{FUN}{a transformer function to apply to the incoming coordinates. As a
convenience if \code{FUN} is missing and \code{xyz} is a function then that
function will be applied to coordinates on the clipboard.}

\item{write_clip}{Whether to write the result to the clipboard. When missing
(the default) will write to clipboard only if coordinates were read from
the clipboard because \code{xyz=NULL}.}
}
\value{
Character vector tab separated coordinates. When \code{xyz} is
  missing these will be returned invisibly and also written to the clipboard.
}
\description{
\code{tabify_coords} can be used to convert a comma separated
  value on the clipboard e.g. from neuroglancer to tab separated values
  needed by Seatable when coordinates are stored in 3 separate columns. It
  also words for multiple coordinates.
}
\examples{
\donttest{
tabify_coords(1:3)
}
\dontrun{
# copy position from clipboard and write back as TSV
tabify_coords()
# same but convert from raw coordinates to nm
tabify_coords(flywire_raw2nm)
}
}
