% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{voxdims.ngscene}
\alias{voxdims.ngscene}
\title{Extract voxel dimensions from a neuroglancer scene}
\usage{
\method{voxdims}{ngscene}(x, units = c("nm", "um", "microns", "mm", "m"), ...)
}
\arguments{
\item{x}{A neuroglancer scene as generated by \code{\link{ngl_decode_scene}}.}

\item{units}{Optionally specify output units for the voxel dimensions.
\code{NA} uses the original units.}

\item{...}{additional arguments (currently ignored)}
}
\description{
Extract voxel dimensions from a neuroglancer scene
}
\seealso{
\code{\link[nat]{voxdims}}
}
