% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nginfo.R
\name{write_nginfo}
\alias{write_nginfo}
\title{Read and write neuroglancer annotation info json files}
\usage{
write_nginfo(anndf, f, sep = "_")
}
\arguments{
\item{anndf}{A data.frame in which the first column contains ids for each
neuron and additional columns contain annotations that will be joined into
a single string.}

\item{f}{Path to a json file. \code{write_nginfo} will create the enclosing
directory if necessary.}

\item{sep}{The separator used to paste multiple columns of annotations
together.}
}
\value{
For \code{read_nginfo} a list containing annotations.For
  \code{write_nginfo}, the path \code{f} invisibly.
}
\description{
Read and write neuroglancer annotation info json files
}
\details{
Note that there is nothing specific to flywire about these two
  functions - they could be used for any data source.
}
\examples{
\donttest{
tf=tempfile(pattern = "info")
df=data.frame(id=c(10000,10002), type=c("DNp01"))
write_nginfo(df, tf)
}
}
\seealso{
\code{\link{fct2nginfo}}
}
