% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng-mesh.R
\name{xyzmatrix.ng_raw}
\alias{xyzmatrix.ng_raw}
\alias{xyzmatrix}
\alias{xyzmatrix.ng_raw_list}
\title{Extract the XYZ coordinates from neuroglancer mesh objects}
\usage{
\method{xyzmatrix}{ng_raw}(x, ...)

\method{xyzmatrix}{ng_raw_list}(x, ...)
}
\description{
Extract the XYZ coordinates from neuroglancer mesh objects
}
\seealso{
\code{\link[nat]{xyzmatrix}}
}
