% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{add_celltype_info}
\alias{add_celltype_info}
\alias{flytable_meta}
\title{Fetch flytable cell type information to a dataframe with flywire ids}
\usage{
add_celltype_info(x, idcol = NULL, version = NULL, ...)

flytable_meta(ids = NULL, version = NULL, ...)
}
\arguments{
\item{x}{a data.frame containing root ids}

\item{idcol}{Optional character vector specifying the column containing ids
of the neurons for which cell type information should be provided.}

\item{version}{Optional numeric CAVE version (see \code{flywire_cave_query}).
The special signalling value of \code{TRUE} uses the current default data
dump as returned by \code{\link{flywire_connectome_data_version}}.}

\item{...}{additional arguments passed to \code{flytable_cell_types}}

\item{ids}{Flywire identifiers/query in any form understood by
\code{\link{flywire_ids}}}
}
\value{
a data.frame with extra columns
}
\description{
\code{add_celltype_info} will add information to an existing
  dataframe.

\code{flytable_meta} will fetch a data.frame of metadata from
  flytable for a given set of identifiers.
}
\details{
the root ids must be in a column called one of \code{"pre_id",
  "post_id", "root_id", "post_pt_root_id", "pre_pt_root_id"}. If you do not
  have exactly one of these columns present then you must specify your
  preferred column with the \code{idcol} argument.
}
\examples{
\donttest{
kcin=flywire_partner_summary("720575940626474889", partners = 'in',
  cleft.threshold = 50)
kcin
kcin2=add_celltype_info(kcin)
kcin2
library(dplyr)
kcin2 \%>\%
  group_by(cell_type) \%>\%
  summarise(wt = sum(weight),n=n()) \%>\%
  arrange(desc(wt))
kcin2 \%>\%
  count(cell_class, wt = weight)
}
\donttest{
flytable_meta("class:MBON")
flytable_meta("type:MBON2\%")
# the / introduces a regex query (small performance penalty, more flexible)
flytable_meta("/type:MBON2[0-5]")
}
}
\seealso{
\code{\link{flytable_cell_types}}
}
