% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_fetch}
\alias{brainmaps_fetch}
\title{GET/POST from brainmaps API}
\usage{
brainmaps_fetch(url, body = NULL, parse.json = TRUE,
  include_headers = FALSE, simplifyVector = TRUE, ...)
}
\arguments{
\item{url}{Full URL for brainmaps API endpoint}

\item{body}{an R list with parameters that will be converted with
\code{jsonlite::\link{toJSON}} and then passed on to \code{\link{POST}}.
You can also pass a \code{JSON} character vector to have more control of
the \code{JSON} encoding.}

\item{parse.json}{Whether or not to parse a JSON response to an R object 
(default \code{TRUE})}

\item{include_headers}{Whether to include basic headers from the http request
as attributes on the parsed JSON object (default \code{TRUE}) when 
\code{parse.json=TRUE}.}

\item{simplifyVector}{Whether to use \code{jsonlite::simplifyVector}}

\item{...}{Additional arguments passed to the \code{httr::GET} or 
\code{httr::POST} function}
}
\value{
An R list parse
}
\description{
GET/POST from brainmaps API
}
\examples{
\dontrun{
brainmaps_fetch("https://brainmaps.googleapis.com/v1/volumes")
}
}
