% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_fetch}
\alias{brainmaps_fetch}
\alias{brainmaps_clear_cache}
\title{GET/POST from brainmaps API with optional retry / cache}
\usage{
brainmaps_fetch(
  url,
  body = NULL,
  parse.json = TRUE,
  cache = FALSE,
  retry = 0L,
  include_headers = FALSE,
  simplifyVector = TRUE,
  ...
)

brainmaps_clear_cache()
}
\arguments{
\item{url}{Full URL for brainmaps API endpoint}

\item{body}{an R list with parameters that will be converted with
\code{jsonlite::\link{toJSON}} and then passed on to \code{\link{POST}}.
You can also pass a \code{JSON} character vector to have more control of
the \code{JSON} encoding.}

\item{parse.json}{Whether or not to parse a JSON response to an R object 
(default \code{TRUE})}

\item{cache}{Whether or not to cache responses (default \code{FALSE})}

\item{retry}{The number of times to retry the operation (default 0,
\code{FALSE}=>\code{0} and \code{TRUE}=>3). See the documentation of the
\code{times} argument of \code{httr::\link{RETRY}} for further details.}

\item{include_headers}{Whether to include basic headers from the http request
as attributes on the parsed JSON object (default \code{TRUE}) when 
\code{parse.json=TRUE}.}

\item{simplifyVector}{Whether to use \code{jsonlite::simplifyVector}}

\item{...}{additional arguments passed to the \code{httr::{RETRY}} function.
This may include a \code{\link[httr]{config}} list other named parameters
etc.}
}
\value{
An R list parse
}
\description{
\code{brainmaps_fetch} calls the brainmaps API with optional
  request cache and retries.

\code{brainmaps_clear_cache} clears the cache used by \code{brainmaps_fetch}
}
\examples{
\dontrun{
brainmaps_fetch("https://brainmaps.googleapis.com/v1/volumes")
# retry up to 4 times on failure
brainmaps_fetch("https://brainmaps.googleapis.com/v1/volumes", retry=4)
# cache results
brainmaps_fetch("https://brainmaps.googleapis.com/v1/volumes", cache=TRUE)
# use arbitrary curl/httr options (see httr_options())
httr::with_config(httr::verbose(), {
  brainmaps_fetch("https://brainmaps.googleapis.com/v1/volumes")
})
}
\donttest{
brainmaps_clear_cache()
}
}
