% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_skeleton}
\alias{brainmaps_skeleton}
\title{Low level function to fetch skeleton from brainmaps API}
\usage{
brainmaps_skeleton(x, skeletonuri = getOption("fafbseg.skeletonuri"), ...)
}
\arguments{
\item{x}{A single segment id}

\item{skeletonuri}{The brainmaps URI describing the skeleton source. Defaults
to the value of \code{options("fafbseg.skeletonuri")}.}

\item{...}{Additional arguments passed to \code{\link{brainmaps_fetch}}}
}
\value{
A list containing the following fields \itemize{

  \item{\code{nvertices}}{ The number of vertices (n)}

  \item{\code{nedges}}{ The number of edges (m)}

  \item{\code{vertices}}{ A \code{n} x 3 matrix of vertex locations}

  }
}
\description{
Low level function to fetch skeleton from brainmaps API
}
\details{
This API seems to return skeletons in physical (i.e. already
  calibrated) coordinates.
}
\examples{
\dontrun{
brainmaps_skeleton(9208128833)
brainmaps_skeleton(9208128833,
  skeletonuri=paste0("brainmaps://772153499790:fafb_v14:",
  "fafb_v14_16nm_v00c_split3xfill2_flatreseg2_skeletons32",
  "/teasar512_nnconn165_mc10000_prune10_thresh1000_sparse250"))
}
}
