% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainmaps-api.R
\name{brainmaps_xyz2id}
\alias{brainmaps_xyz2id}
\title{Convert 3D x,y,z locations in brainmaps volumes to segmentation ids}
\usage{
brainmaps_xyz2id(xyz,
  volume = "772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2",
  voxdims = c(8, 8, 40), ...)
}
\arguments{
\item{xyz}{Nx3 matrix of points or an object containing vertex data that is
compatible with \code{\link{xyzmatrix}}. These should be in physical space
(i.e. nm) unless \code{voxdims=NULL}.}

\item{volume}{character vector identifier string for the volume containing
segmentation data - see examples}

\item{voxdims}{the implied voxel dimensions for the volume. If set to
\code{NULL} then the function will not attempt to scale the incoming x,y,z
locations.}

\item{...}{Additional arguments passed to \code{\link{brainmaps_fetch}}}
}
\value{
A numeric vector of Google segment ids
}
\description{
Convert 3D x,y,z locations in brainmaps volumes to segmentation ids
}
\examples{
\dontrun{
brainmaps_xyz2id(c(54171, 21026,3212), voxdims = NULL)
brainmaps_xyz2id(c(433368, 168208, 128480))

library(elmr)
# get a manually traced neuron (just keep first and only entry in neuronlist)
dl4=read.neurons.catmaid('glomerulus DL4 right')[[1]]
# map every node location to segmentation ids
dl4.segs=brainmaps_xyz2id(dl4)
# remove unmapped locations which get id 0
dl4.segs=setdiff(dl4.segs, 0)
# read in corresponding skeletons
dl4.skels=read_segments2(dl4.segs)
# read in corresponding skeletons after including agglomeration merge groups
dl4.allskels=read_segments2(find_merged_segments(dl4.segs))
}
}
