% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose-segmentation.R
\name{choose_segmentation}
\alias{choose_segmentation}
\alias{with_segmentation}
\title{Choose or (temporarily) use a FAFB autosegmentation}
\usage{
choose_segmentation(
  release = c("20190805", "20190521", "flywire31", "sandbox-flywire31"),
  set = TRUE
)

with_segmentation(release, expr)
}
\arguments{
\item{release}{character vector specifying a released segmentation via a
known short name or a sample neuroglancer URL.}

\item{set}{Whether or not to set the selected options for the selected
\code{release}.}

\item{expr}{The expression to execute with the temporary options set}
}
\value{
If \code{set=TRUE} a list containing the previous values of the
  relevant global options (in the style of \code{\link{options}}. If
  \code{set=FALSE} a named list containing the option values.

The result of evaluating \code{expr}
}
\description{
\code{with_segmentation} allows a specific segmentation to be
  temporarily selected.
}
\details{
Each released segmentation implies a number of global options. This
  package comes with 4 different default scene urls specified via
  \code{\link{choose_segmentation}} or \code{\link{with_segmentation}}. This
  is the easiest way to choose a particular segmentation. You can also pass a
  sample URL to the \code{release} argument.

  You can also specify a different sample URL via the \code{sampleurl}
  argument of some functions; it will be remembered for the rest of the R
  session. If you regularly use a particular kind of scene URL, you can set
  \code{options(fafbseg.sampleurl)} in your \code{\link{Rprofile}} file.

  If you need to use both built-in and custom segmentation URLs, we recommend
  specifying the custom URL in your \code{\link{Rprofile}} file and using the
  \code{with_segmentation} to run code that uses one of the built-in
  segmentations.
}
\examples{
\donttest{
choose_segmentation('20190805', set=FALSE)
}
\dontrun{
# temporarily change default segmentation to run a command
# but restore original default when finished
with_segmentation('flywire31', {open_fafb_ngl(c(460792, 221812, 61480))})
}
\dontrun{
n <- with_segmentation("20190521",
  read.neuron.brainmaps(22427007374))

# open location in flywire
with_segmentation("flywire", open_fafb_ngl(c(433440, 168344, 131200)))
}
}
