% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-skeleton.R
\name{fafb14_to_flywire_ids}
\alias{fafb14_to_flywire_ids}
\title{Get flywire IDs that map onto CATMAID neurons}
\usage{
fafb14_to_flywire_ids(
  search,
  pid = 1L,
  conn = NULL,
  fetch.annotations = FALSE,
  ...
)
}
\arguments{
\item{search}{one or more skids or a CATMAID query expression. Else, a
neuronlist of neurons in FAFB14 space.}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link[catmaid]{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link[catmaid]{catmaid_login}}.}

\item{fetch.annotations}{Whether or not to fetch the annotations for each
skeleton (default \code{FALSE})}

\item{...}{Additional arguments passed to \code{nat::nlapply}.}
}
\value{
A \code{data.frame} of ranked flywire IDs, and the number of points
  in each CATMAID neuron that maps to that ID.
}
\description{
Provide this function with a CATMAID query (skeleton IDs or an
  annotation term, as you could provide to \code{catmaid::catmaid_skids}) or
  a \code{nat::neuronlist} object in FAFB14 space, and it will return a
  ranked list of flywire IDs that map onto the given neuron(s).
}
\examples{
\dontrun{
df = fafb14_to_flywire_ids("16")
head(df)
}
}
