% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fafbseg-ids}
\alias{fafbseg-ids}
\alias{swc2segmentid}
\alias{segmentid2zip}
\alias{zip2segmentstem}
\title{Convert between filenames and neuroglancer ids}
\usage{
swc2segmentid(x, include.fragment = FALSE)

segmentid2zip(x)

zip2segmentstem(x)
}
\arguments{
\item{x}{Input file or id}

\item{include.fragment}{Whether to include the sub identifier of the skeleton
fragment (see details).}
}
\value{
for \code{swc2segmentid} a numeric vector or matrix depending on the
  value of \code{include.fragment}
}
\description{
\code{swc2segmentid} converts an swc filename to a segment id

\code{segmentid2zip} converts a segment id to the zip file that
  contains it

\code{zip2segmentstem} converts a zip file to the initial part
  of the segment id i.e. the segment stem (see details).
}
\details{
Segment ids are unique integers. There are about 8E8 in the current
  skeletonisation but it seems that the ids can still be > 2^31 (usually
  \code{.Machine$integer.max}). Therefore they will be stored in R as numeric
  values or the \code{bit64::integer64} values.

  Each segmentation has keen skeletonised however this usually results in
  multiple skeleton fragments which have been written out as separate SWC
  files: \code{"named <segment id>.<fragment>.swc"}

  Each segment id is mapped onto a zip file by dividing by 1e5 and discarding
  the remainder.
}
\examples{
swc2segmentid("10001654273.1.swc")
swc2segmentid(sprintf("10001654273.\%d.swc", 0:2), include.fragment=TRUE)
segmentid2zip(10001654273)
segmentid2zip(swc2segmentid("10001654273.1.swc"))
}
