% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg-package.R
\docType{package}
\name{fafbseg-package}
\alias{fafbseg}
\alias{fafbseg-package}
\title{fafbseg: Support functions for analysis of FAFB EM segmentation}
\description{
Functions that read some of the raw data formats produced by Google
electron microscopy segmentation tools as displayed by 'neuroglancer' and then
allow basic 3D visualisation and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/fafbseg}
  \item Report bugs at \url{https://github.com/jefferis/fafbseg/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
