% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg-package.R
\docType{package}
\name{fafbseg-package}
\alias{fafbseg}
\alias{fafbseg-package}
\title{fafbseg: Support functions for analysis of FAFB EM segmentation}
\description{
Functions that read some of the raw data formats produced by Google
  electron microscopy segmentation tools as displayed by 'neuroglancer' and then
  allow basic 3D visualisation and analysis.
}
\section{Package Options}{
 \itemize{

  \item{\code{fafbseg.skelziproot}} set to the location of a folder
  containing the zipped versions of the skeletonised segmentations. This will
  be used by \code{\link{read_segments}}, \code{\link{read_topn}} etc.

  \item{\code{fafbseg.skeletonuri}} a brainmaps URI specifying a remote
  source used by \code{\link{read.neurons.brainmaps}} and
  \code{\link{brainmaps_skeleton}} to read neuronal skeletons.

  \item{\code{fafbseg.sampleurl}} optionally set to a sample Neuroglancer URL
  that will modified to point to arbitrary locations by
  \code{\link{open_fafb_ngl}}.

  \item{\code{fafbseg.brainmaps_xyz2id.chunksize}} this will default to
  querying 4000 vertices at a time. Set this smaller if the queries time out
  or larger to speed things up. See

  }
}

\examples{
options()[grep("^fafbseg\\\\.", names(options()))]
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/fafbseg}
  \item Report bugs at \url{https://github.com/jefferis/fafbseg/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
