% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-all.R
\name{fetch_all_curl}
\alias{fetch_all_curl}
\title{Fetch data specified by a set of curl shell comands (e.g. from Chrome)}
\usage{
fetch_all_curl(x, outdir = NULL, regex = "^curl", ...)
}
\arguments{
\item{x}{Path to a file or character vector of statements}

\item{outdir}{Optional output directory (will be created if necessary)}

\item{regex}{Optional regular expression that curl statements must match}

\item{...}{Additional arguments to \code{\link{grep}}}
}
\value{
The path to the output directory invisibly
}
\description{
Fetch data specified by a set of curl shell comands (e.g. from Chrome)
}
\examples{
\dontrun{
fetch_all_curl("all_curl.sh", outdir="alldata", regex="meshes:batch", fixed=TRUE)
}
}
