% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-groups.R
\name{find_merged_segments}
\alias{find_merged_segments}
\title{Find all merged segment ids for a given starting google segment id}
\usage{
find_merged_segments(x, return.groups = FALSE)
}
\arguments{
\item{x}{a segment id or any other input that can be interpreted by
\code{\link{ngl_segments}}}

\item{return.groups}{Whether to return the merge groups as well as the
segment ids}
}
\value{
vector of segment ids (in ascending order)
}
\description{
Find all merged segment ids for a given starting google segment id
}
\details{
segment ids in \code{ffn16reseg-ms1000_md0.02_c0.6_iou0.7} always
  match one raw segment id in \code{fafb_v14_16nm_v00c_split3xfill2} but may
  map to a large agglomerated merge group.

  Note that this functions depends on the
  \code{\link[fafbsegdata]{segment_merge_groups.dt}} object from the
  \code{\link[fafbsegdata]{fafbsegdata}} package.
}
\examples{
\donttest{
find_merged_segments(7186840767)
# these all belong to one merge group by definition
find_merged_segments(7186840767, return.groups=TRUE)
}
}
