% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable-queries}
\alias{flytable-queries}
\alias{flytable_list_rows}
\alias{flytable_query}
\title{Flytable database queries}
\usage{
flytable_list_rows(
  table,
  base = NULL,
  view_name = NULL,
  order_by = NULL,
  desc = FALSE,
  start = 0L,
  limit = Inf,
  python = FALSE,
  chunksize = NULL
)

flytable_query(
  sql,
  limit = 100000L,
  base = NULL,
  python = FALSE,
  convert = TRUE
)
}
\arguments{
\item{table}{The name of a table inside your database}

\item{base}{Character vector naming a seatable base (recommended) or a
\code{Base} object returned by \code{flytable_base} (expert use).}

\item{view_name}{An optional view which may limit the rows/columns displayed.}

\item{order_by}{Optional name of columns to order results}

\item{desc}{Whether to use descending order (default \code{FALSE} =>
ascending order)}

\item{start}{Optional starting row}

\item{limit}{An optional limit, which only applies if you do not specify a
limit directly in the \code{sql} query. By default seatable limits SQL
queries to 100 rows. We increase the limit to 100000 rows by default.}

\item{python}{Whether to return a Python pandas \code{DataFrame}. The default
of \code{FALSE} returns an R \code{data.frame}}

\item{chunksize}{Optional The maximum number of rows to request in one web
request. For advanced use only as the default value of \code{NULL} will
fetch as many as possible.}

\item{sql}{A SQL query string. See examples and
\href{https://seatable.github.io/seatable-scripts/python/query/}{seatable
docs}.}

\item{convert}{Expert use only: Whether or not to allow the Python seatable
module to process raw output from the database. This is is principally for
debugging purposes. NB this imposes a requirement of seatable_api >=2.4.0.}
}
\value{
An R \code{data.frame} or Pandas \code{DataFrame} depending on the
  value of the \code{python} argument.

a \code{data.frame} of results. There should be 0 rows if no rows
  matched query.
}
\description{
\code{flytable_query} performs a SQL query against a flytable
  database. You can omit the \code{base} argument unless you have tables of
  the same name in different bases.
}
\details{
Flytable uses programmatic access to the
  \href{https://seatable.github.io/seatable-scripts/}{seatable} API.
}
\examples{
\donttest{
flytable_list_rows(table = "testfruit")
}
\donttest{
flytable_query("SELECT person, fruit_name FROM testfruit WHERE person!='Bob'")
}
\dontrun{
flytable_query(paste("SELECT root_id, supervoxel_id FROM info limit 5"))
}
}
\seealso{
\code{\link{tabify_coords}} to help with copy-pasting coordinates to
  seatable.

Other flytable: 
\code{\link{flytable_list_selected}()},
\code{\link{flytable_login}()},
\code{\link{flytable_update_rows}()}
}
\concept{flytable}
