% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable-queries}
\alias{flytable-queries}
\alias{flytable_list_rows}
\alias{flytable_query}
\title{Flytable database queries}
\usage{
flytable_list_rows(
  base,
  table,
  view_name = NULL,
  order_by = NULL,
  desc = FALSE,
  start = NULL,
  limit = Inf,
  python = FALSE
)

flytable_query(base, sql, python = FALSE)
}
\arguments{
\item{base}{Character vector naming a seatable base (recommended) or a
\code{Base} object returned by \code{flytable_base} (expert use).}

\item{table}{The name of a table inside your database}

\item{view_name}{An optional view which may limit the rows/columns displayed.}

\item{order_by}{Optional name of columns to order results}

\item{desc}{Whether to use descending order (default \code{FALSE} =>
ascending order)}

\item{start}{Optional starting row}

\item{limit}{Maximum number of rows to return (the default \code{Inf} implies
all rows)}

\item{python}{Whether to return a Python pandas \code{DataFrame}. The default
of \code{FALSE} returns an R \code{data.frame}}

\item{sql}{A SQL query string. See examples and
\href{https://seatable.github.io/seatable-scripts/python/query/}{seatable
docs}.}
}
\value{
An R \code{data.frame} or Pandas \code{DataFrame} depending on the
  value of the \code{python} argument.


}
\description{
\code{flytable_query} performs a SQL query against a flytable database.
}
\details{
Flytable uses programmatic access to the
  \href{https://seatable.github.io/seatable-scripts/}{seatable} API.
}
\examples{
\donttest{
flytable_list_rows("hemilineages", "fruit")
}
\donttest{
flytable_query("hemilineages", "SELECT person, fruit_name FROM testfruit WHERE person!='Bob'")
}
\dontrun{
flytable_query("hemilineages", "select FLYWIREsvid, hemibrain_match FROM fafb_hemilineages_survey WHERE hemibrain_match!='' limit 5")
}
}
\seealso{
Other flytable: 
\code{\link{flytable_login}()}
}
\concept{flytable}
