% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable_cell_types}
\alias{flytable_cell_types}
\title{Fetch (memoised) flywire cell type information from flytable}
\usage{
flytable_cell_types(
  pattern = NULL,
  version = NULL,
  timestamp = NULL,
  target = c("type", "cell_type", "hemibrain_type", "cell_class", "all"),
  transfer_hemibrain_type = c("extra", "none", "all"),
  cache = TRUE
)
}
\arguments{
\item{pattern}{Optional character vector specifying a pattern that cell types
must match in a SQL \code{LIKE} statement executed by
\code{\link{flytable_query}}. The suffix \code{_L} or \code{_R} can be used
to restricted to neurons annotated to the L or R hemisphere. See examples.}

\item{version}{An optional CAVE materialisation version
number. See details and examples.}

\item{timestamp}{An optional timestamp as a string or POSIXct, interpreted as
UTC when no timezone is specified.}

\item{target}{A character vector specifying which flytable columns
\code{pattern} should match. The special value of \code{type} means either
\code{cell_type} \emph{or} \code{hemibrain_type} should match. The special value of \code{all} means to match against any of \code{cell_type, hemibrain_type, cell_class}.}

\item{transfer_hemibrain_type}{Whether to transfer the \code{hemibrain_type}
column into the \code{cell_type} (default TRUE, see details)}

\item{cache}{Whether to cache the results for 5m (default \code{TRUE} since
this is a little expensive)}
}
\value{
The original data.frame left joined to appropriate rows from
  flytable.
}
\description{
Fetch (memoised) flywire cell type information from flytable
}
\details{
when \code{transfer_hemibrain_type=TRUE}, \code{hemibrain_type}
  values will be transferred into the \code{cell_type} column if
  \code{cell_type} is empty.
}
\examples{
\donttest{
flytable_cell_types("MBON\%")
flytable_cell_types("MBON\%", version=450)
# two characters
flytable_cell_types("MBON__")
# at least one character
flytable_cell_types("MBON_\%")
# range
flytable_cell_types("MBON2[0-5]")

# include side specification
flytable_cell_types("DA2_lPN_R")
# only the RHS MBON20
flytable_cell_types("MBON20_R")
# all RHS cells with class MBON
flytable_cell_types("MBON_R", target="cell_class")

# anything with type *OR* class information
cells=flytable_cell_types(target = 'all')
# anything that mentions PN anywhere
pncands=flytable_cell_types('\%PN\%', target = 'all')
}
}
\seealso{
\code{\link{add_celltype_info}}
}
