% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable_list_selected}
\alias{flytable_list_selected}
\title{List selected rows from flytable}
\usage{
flytable_list_selected(
  ids = NULL,
  table = "info",
  fields = "*",
  idfield = "root_id",
  ...
)
}
\arguments{
\item{ids}{One or more identifiers}

\item{table}{The name of the flytable table}

\item{fields}{The database columns to return}

\item{idfield}{Which field to use as a key for lookup}

\item{...}{Additional arguments passed to \code{\link{flytable_query}}}
}
\value{
a dataframe containing the selected rows / columns
}
\description{
List selected rows from flytable
}
\seealso{
Other flytable: 
\code{\link{flytable-queries}},
\code{\link{flytable_login}()},
\code{\link{flytable_update_rows}()}
}
\concept{flytable}
