% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable_login}
\alias{flytable_login}
\alias{flytable_base}
\title{Low level functions to access the flytable metadata service}
\usage{
flytable_login(
  url = "https://flytable.mrc-lmb.cam.ac.uk/",
  user = Sys.getenv("FLYTABLE_USER"),
  pwd = Sys.getenv("FLYTABLE_PWD")
)

flytable_base(
  base_name,
  workspace_id = NULL,
  url = "https://flytable.mrc-lmb.cam.ac.uk/",
  cached = FALSE
)
}
\arguments{
\item{url}{Optional URL to the server}

\item{user, pwd}{flytable user and password, normally retrieved from
\code{FLYTABLE_USER, FLYTABLE_PWD} environment variables.}

\item{base_name}{Character vector specfying the \code{base}}

\item{workspace_id}{A numeric id specifying the workspace. Advanced use only
since we can normally figure this out from \code{base_name}.}

\item{cached}{Whether to use a cached base object}
}
\value{
For \code{flytable_login}, a Python
  \href{https://seatable.github.io/seatable-scripts/python/account/}{\code{Account}}
   object from the seatable api as wrapped by reticulate.

For \code{flytable_base}, a Python
  \href{https://seatable.github.io/seatable-scripts/python/base/}{\code{Base}}
   object from the seatable api as wrapped by reticulate.
}
\description{
\code{flytable_login} uses your flytable user name and email to
  log into the service.

\code{flytable_base} returns a \code{base} object (equivalent to
  a mysql database) which allows you to access one or more tables, logging in
  to the service if necessary.
}
\details{
Normally you should not need to use these lower level functions
  directly, instead preferring higher level functions such as
  \code{\link{flytable_query}}.

  In order to start using flytable, you must set user and password
  environment variables. This will normally be in your .Renviron file e.g. by
  doing \code{usethis::edit_r_environ()} and then setting \code{
  FLYTABLE_USER="<email>" FLYTABLE_PWD="xxxx"} always making sure that you
  have a blank line at the end of the file.

\code{flytable_base} will use your flytable user name and email to
  log into the service.
}
\examples{
\dontrun{
flytable_login()
flytable_login(user='alice@gmail.com', pwd='xxxx')
}
\dontrun{
hemilineages=flytable_base('hemilineages')
}

}
\seealso{
Other flytable: 
\code{\link{flytable-queries}}
}
\concept{flytable}
