% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable_login}
\alias{flytable_login}
\alias{flytable_set_token}
\alias{flytable_base}
\alias{flytable_alltables}
\alias{flytable_columns}
\title{Low level functions to access the flytable metadata service}
\usage{
flytable_login(
  url = "https://flytable.mrc-lmb.cam.ac.uk/",
  token = Sys.getenv("FLYTABLE_TOKEN", unset = NA_character_)
)

flytable_set_token(user, pwd, url = "https://flytable.mrc-lmb.cam.ac.uk/")

flytable_base(
  table = NULL,
  base_name = NULL,
  workspace_id = NULL,
  url = "https://flytable.mrc-lmb.cam.ac.uk/",
  cached = TRUE
)

flytable_alltables(ac = NULL, cached = TRUE)

flytable_columns(table, base = NULL, cached = TRUE)
}
\arguments{
\item{url}{Optional URL to the server}

\item{token}{normally retrieved from \code{FLYTABLE_TOKEN} environment
variable.}

\item{user, pwd}{flytable user and password used by \code{flytable_set_token}
to obtain a token}

\item{table}{Character vector specifying a table foe which you want a
\code{base} object.}

\item{base_name}{Character vector specifying the \code{base}}

\item{workspace_id}{A numeric id specifying the workspace. Advanced use only
since we can normally figure this out from \code{base_name}.}

\item{cached}{Whether to use a cached version of the response if available.
Set to \code{FALSE} if you know tables have been added or renamed during
your session.}

\item{ac}{Optional account object returned by flytables_login}

\item{base}{Optional character vector naming a seatable base (recommended) or
a \code{Base} object returned by \code{\link{flytable_base}} (expert use).
The default value of \code{NULL} will rely on the \code{table} so long as
it is unique across the flytable server.}
}
\value{
For \code{flytable_login}, a Python
  \href{https://seatable.github.io/seatable-scripts/python/account/}{\code{Account}}
   object from the seatable api as wrapped by reticulate.

For \code{flytable_base}, a Python
  \href{https://seatable.github.io/seatable-scripts/python/base/}{\code{Base}}
   object from the seatable api as wrapped by reticulate.

A \code{data.frame} containing the \code{base_name},
  \code{workspace_id}, table \code{name} and table \code{_id}.

\code{flytable_columns} a data.frame containing columns \code{name},
  \code{type} and \code{rtype}
}
\description{
\code{flytable_login} uses your flytable user name and email to
  log into the service.

\code{flytable_set_token} will obtain and store a permanent
  seatable user-level API token.

\code{flytable_base} returns a \code{base} object (equivalent to
  a mysql database) which allows you to access one or more tables, logging in
  to the service if necessary. The returned base object give you full access
  to the Python
  \href{https://seatable.github.io/seatable-scripts/python/base/}{\code{Base}}
   API allowing a range of row/column manipulations.

\code{flytable_alltables} lists all tables across all flytables
  bases.

\code{flytable_columns} returns the name and type of all regular
  columns in a base as well as the default R type. Private columns such as
  \code{_id} are not included.
}
\details{
Besides initial setup (next paragraph), you should not need to use
  these lower level functions directly. Instead we recommend higher level
  functions such as \code{\link{flytable_query}}.

  In order to start using flytable, you must get an API token. There doesn't
  seem to be a convenient way to do this from the seatable web interface but
  you can get one by calling \code{flytable_set_token} with your flytable
  user and password. This should be a once only step. Thereafter you should
  have a \code{FLYTABLE_TOKEN} environment variable set in your .Renviron
  file.

\code{flytable_base} will use your flytable API token to log into
  the service.
}
\examples{
\dontrun{
flytable_login()
}
\dontrun{
flytable_set_token(user='xxx@gmail.com', pwd='yyy')
}
\dontrun{
hemilineages=flytable_base(base_name='hemilineages')
# equivalent, but simpler since you only have to remember the table name
hemilineages=flytable_base('fafb_hemilineages_survey')
}

\donttest{
flytable_alltables()
}
\donttest{
flytable_columns("info")
}
}
\seealso{
Other flytable: 
\code{\link{flytable-queries}},
\code{\link{flytable_update_rows}()}
}
\concept{flytable}
