% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable-cache.R
\name{flytable_now}
\alias{flytable_now}
\title{Fetch current server time from flytable}
\usage{
flytable_now(table = "info")
}
\arguments{
\item{table}{Table name to query (used to verify connectivity)}
}
\value{
Character timestamp from server
}
\description{
Fetch current server time from flytable
}
\keyword{internal}
