% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytable.R
\name{flytable_update_rows}
\alias{flytable_update_rows}
\title{Update rows in a flytable database}
\usage{
flytable_update_rows(df, table, base = NULL, chunksize = 1000L, ...)
}
\arguments{
\item{df}{A data.frame containing the data to upload including an \code{_id}
column that can identify each row in the remote table.}

\item{table}{Character vector naming a table}

\item{base}{Character vector naming a seatable base (recommended) or a
\code{Base} object returned by \code{flytable_base} (expert use).}

\item{chunksize}{To split large requests into smaller ones with max this many
rows.}

\item{...}{Additional arguments passed to \code{\link[pbapply]{pbsapply}}
which might include \code{cl=2} to specify a number of parallel jobs to
run.}
}
\value{
Logical indicating success, invisibly (failures will normally cause
  premature termination with errors written to the console).
}
\description{
Update rows in a flytable database
}
\details{
seatable automatically maintains a unique id for each row in a
  \code{_id} column. This is returned by flytable_query and friends. If you
  modify data and then want to update again, you need to keep the column
  containing this row \code{_id}.

  The \code{chunksize} argument is required because it seems that there is a
  maximum of 1000 rows per update action.
}
\examples{
\donttest{
fruit=flytable_list_rows('testfruit')
flytable_update_rows(table='testfruit', fruit[c(1,4:6)])
}
}
\seealso{
Other flytable: 
\code{\link{flytable-queries}},
\code{\link{flytable_login}()}
}
\concept{flytable}
