% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire.R
\name{flywire2fafb}
\alias{flywire2fafb}
\title{Map points in  FlyWire v1 TO FAFB14 space (xyz nm)}
\usage{
flywire2fafb(
  xyz,
  method = c("mapmany", "map1"),
  chunksize = 200,
  swap = FALSE,
  ...
)
}
\arguments{
\item{xyz}{A Nx3 matrix of points}

\item{method}{Whether to map many points at once (default) or just one}

\item{chunksize}{The number of points to send to the server when mapping many
points at once.}

\item{...}{Additional arguments passed to httr::GET/POST operation}
}
\value{
an Nx3 matrix of points
}
\description{
Map points in  FlyWire v1 TO FAFB14 space (xyz nm)
}
\details{
Note that you can also access FlyWire->FAFB bridging registration
  via the \code{\link{xform_brain}} series of functions. This will allow you
  to transform most kinds of 3D data objects, whereas the \code{flywire2fafb}
  function is restricted to plain 3D coordinates. See examples.

  Mapping single points is unlikely to be useful, but you may wish to adjust
  the \code{chunksize} argument to send more points at once at the risk of
  possible server timeouts. The value of 200 is quite conservative.
}
\examples{
flywire2fafb(cbind(163826*4, 75263*4, 3513*40))
flywire2fafb(cbind(163826*4, 75263*4, 3513*40), method="map1")

data("AV4b1", package='catmaid')
set.seed(42)
before=xyzmatrix(AV4b1)[sample(nvertices(AV4b1), size=2000), ]
after=flywire2fafb(before)
d=sqrt(rowSums((before-after)^2))
hist(d, br=20)

\dontrun{
AV4b1.flywire <- xform_brain(AV4b1, reference="FlyWire", sample="FAFB14")
plot3d(neuronlist(AV4b1.flywire, AV4b1))
}
}
