% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{flywire_cave_client}
\alias{flywire_cave_client}
\title{Low level access to the Flywire CAVE annotation system}
\usage{
flywire_cave_client(
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production")
)
}
\arguments{
\item{datastack_name}{defaults to the value selected by
\code{\link{choose_segmentation}} and to "flywire_fafb_production" when
that is missing. See \url{https://global.daf-apis.com/info/} for other
options.}
}
\value{
The \code{caveclient.frameworkclient.CAVEclientFull} Python module
  wrapped by reticulate.
}
\description{
Low level access to the Flywire CAVE annotation system
}
\details{
This depends on installation of the Python caveclient library. See
  \code{\link{flywire_cave_query}} for more details.

  This function memoises the initialisation of the Python \code{caveclient}
  once every 12 hours in a given session. Note that on the Python side, the
  client caches the current materialisation version, which typically changes
  every 1-3 days depending on the project. Therefore if you initialise the
  client 2h before a new materialiastion becomes available it will be 10h
  before your client is reinitialised and switches to the new session.

  By default the caveclient logger level is set to \code{"WARNING"}, to
  suppress routine \code{INFO} that might otherwise be captured by R side
  functions. Genuine caveclient \code{WARNING} / \code{ERROR} records still
  propagate. Override via \code{options(fafbseg.caveclient.loglevel =
  "INFO")} (or any other Python \code{logging} level name).
}
\examples{
\donttest{
fac <- flywire_cave_client()
fac$annotation$get_tables()
fac$annotation$get_table_metadata('nuclei_v1')

# annotation tables that have been materialised in order to map XYZ
# points onto current root ids (and immutable supervoxel ids)
# typically the same as fac$annotation$get_tables()
fac$materialize$get_tables()

info=fac$info$get_datastack_info()
# the default synapse table for the dataset
info$synapse_table
}

\dontrun{
# get help on python cave client
reticulate::py_help(fac$info$get_datastack_info)
}
}
