% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{flywire_cave_client}
\alias{flywire_cave_client}
\title{Low level access to the Flywire CAVE annotation system}
\usage{
flywire_cave_client(datastack_name = "flywire_fafb_production")
}
\arguments{
\item{datastack_name}{defaults to "flywire_fafb_production". See
\url{https://global.daf-apis.com/info/} for other options.}
}
\value{

}
\description{
Low level access to the Flywire CAVE annotation system
}
\details{
This depends on installation of the Python caveclient library. You
  can find out more at \url{https://caveclient.readthedocs.io/} as well as
  looking at the Python notebooks on the github repo
  \url{https://github.com/seung-lab/CAVEclient}.

  The annotation system shares authentication infrastructure with the rest of
  the FlyWire API (see \code{\link{flywire_set_token}}).
}
\examples{
\donttest{
fac <- flywire_cave_client()
fac$annotation$get_tables()
fac$annotation$get_table_metadata('nuclei_v1')

# annotation tables that have been materialised in order to map XYZ
# points onto current root ids (and immutable supervoxel ids)
# typically the same as fac$annotation$get_tables()
fac$materialize$get_tables()

info=fac$info$get_datastack_info()
# the default synapse table for the dataset
info$synapse_table
}
}
