% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{flywire_cave_query}
\alias{flywire_cave_query}
\title{Query the FlyWire CAVE annotation system}
\usage{
flywire_cave_query(
  table,
  datastack_name = "flywire_fafb_production",
  live = TRUE,
  ...
)
}
\arguments{
\item{table}{The name of the table to query}

\item{datastack_name}{defaults to "flywire_fafb_production". See
\url{https://global.daf-apis.com/info/} for other options.}

\item{live}{Whether to use live query mode, which updates any root ids to
their current value.}

\item{...}{Additional arguments to the query method. See examples and
details.}
}
\value{
A \code{tibble}. Note that xyzmatrix can be used on single columns
  containing XYZ locations.
}
\description{
Query the FlyWire CAVE annotation system
}
\examples{
\donttest{
# note use of limit to restrict the number of rows
n10=flywire_cave_query(table = 'nuclei_v1', limit=10)
head(as.data.frame(n10))
}
\dontrun{
nuclei_v1=flywire_cave_query(table = 'nuclei_v1')
points3d(xyzmatrix(nuclei_v1$pt_position))

library(elmr)
# calculate signed distance to FAFB surface
# NB this surface is not a perfect fit in the optic lobes
nuclei_v1$d=pointsinside(xyzmatrix(nuclei_v1$pt_position), FAFB.surf,
  rval = 'dist')
points3d(xyzmatrix(nuclei_v1$pt_position),
  col=matlab::jet.colors(20)[cut(nuclei_v1$d,20)])
plot3d(FAFB)
}
}
\seealso{

}
