% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-arrow.R
\name{flywire_connectome_data}
\alias{flywire_connectome_data}
\title{Access precomputed flywire connectivity data}
\usage{
flywire_connectome_data(
  type = c("syn", "pre", "post"),
  version = NULL,
  cached = TRUE,
  ...
)
}
\arguments{
\item{type}{Character vector specifying the kind of data}

\item{version}{Optional CAVE version. The default value of \code{NULL} uses
the latest version available.}

\item{cached}{When version is \code{NULL} whether to use a cached value
(lasting 1 hour) of the latest available version.}

\item{...}{Additional arguments passed to \code{arrow::open_dataset}.}
}
\value{
An arrow object that you can use with \code{dplyr} verbs like
  \code{filter} in order to find neurons/connectivity data of interest.
}
\description{
Access precomputed flywire connectivity data
}
\details{
This depends on precomputed data dumps prepared periodically by Sven
  Dorkenwald. You must either download these to an appropriate location on
  your hard drive or add a link to your own Google drive. I have noticed
  unfortunately that some versions of arrow do not play nicely with the
  Google drive app on macosx, so I would generally recommend downloading a
  specific version to a location on your machine. The link to Sven's Google
  drive folder can be found in
  \href{https://flywire-forum.slack.com/archives/C01M4LP2Y2D/p1644529750249139}{this
   Slack message}.
}
\examples{
\donttest{
# latest available version/
syn=try(flywire_connectome_data('syn'), silent=T)
syn450=try(flywire_connectome_data('syn', version=450), silent=T)
if(!inherits(syn450, 'try-error')) {
syn450
syn450$metadata

dl4ds <- syn450 \%>\%
  filter(pre_pt_root_id==flywire_ids("DL4_adPN_R", version=450, integer64 = T)) \%>\%
  collect()
}

}
}
