% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-arrow.R
\name{flywire_connectome_data}
\alias{flywire_connectome_data}
\title{Access precomputed flywire connectivity data}
\usage{
flywire_connectome_data(
  type = c("syn", "pre", "post"),
  version = NULL,
  cached = TRUE,
  ...
)
}
\arguments{
\item{type}{Character vector specifying the kind of data}

\item{version}{Optional CAVE version. The default value of \code{NULL} uses
the latest version available.}

\item{cached}{When version is \code{NULL} whether to use a cached value
(lasting 1 hour) of the latest available version.}
}
\value{
A
}
\description{
Access precomputed flywire connectivity data
}
\details{
This depends on precomputed data dumps prepared periodically by Sven
  Dorkenwald. You must either download these to an appropriate location on
  your hard drive or add a link to your own google drive. The link to Sven's
  google drive folder can be found in
  \href{https://flywire-forum.slack.com/archives/C01M4LP2Y2D/p1644529750249139}{this
   Slack message}.
}
\examples{
\donttest{
syn450=try(flywire_connectome_data('syn', version=450), silent=T)
if(!inherits(syn450, 'try-error')) {
syn450
syn450$metadata

dl4ds <- syn450 \%>\%
  filter(pre_pt_root_id==flywire_ids("DL4_adPN_R", version=450, integer64 = T)) \%>\%
  collect()

}

}
}
