% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosyn.R
\name{flywire_dcvs}
\alias{flywire_dcvs}
\title{Get predicted dense core vesicle locations in the flywire dataset}
\usage{
flywire_dcvs(
  rootid,
  dataset = c("dcv.2.0", "dcv.1.0"),
  islatest = TRUE,
  simplify = TRUE,
  return = c("parsed", "text", "response"),
  token = NULL,
  simplifyVector = TRUE,
  include_headers = FALSE,
  OmitFailures = TRUE,
  cl = NULL,
  ...
)
}
\arguments{
\item{rootid}{flywire rootid/rootids}

\item{dataset}{which DCV data set from Stephan Gerhard to access. 1.0 is just for the antennal lobes of FAFB. 2.0 is the first run at the whole brain.}

\item{islatest}{logical, whether or not to fetch the latest root_id if the given IDs are out of date.}

\item{simplify}{logical, if \code{FALSE} each rootid is a separate set of two data frames (one for DCV positions, one for the neuron's synapses).
Else, a list of two combined data frames is returned.}

\item{return}{One of "parsed", "text" (for raw JSON), or "response"}

\item{token}{Optional chunkedgraph token (otherwise the default one for the
current segmentation will be used).}

\item{simplifyVector}{Whether to use \code{jsonlite::simplifyVector}}

\item{include_headers}{Whether to include basic headers from the http request
as attributes on the parsed JSON object (default \code{TRUE}) when 
\code{parse.json=TRUE}.}

\item{OmitFailures}{logical, if \code{TRUE} then requests that result in a 500 error are dropped and a warning is displayed
but not an error.}

\item{cl}{A cluster object created by \code{parallel::makeCluster}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations.}

\item{...}{additional arguments passed to the \code{httr::{RETRY}} function.
This may include a \code{\link[httr]{config}} list other named parameters
etc.}
}
\value{
A list, where the first entry contains DCV locations and the second the synapses for the given rootid, with the nearest DCV precalculated for each synapse..
}
\description{
Preliminary dense core vesicle (DCV) detection results from the Lee group, Stephan Gerhard and Minsu Kim.
}
\examples{
\donttest{
# Just AL test data set
data = flywire_dcvs("720575940629166904", dataset = "dcv.1.0")
dcv = data$dcv
synapse = data$syns

# Whole brain data set
dcv = flywire_dcvs(c("720575940631973089","720575940629166904"), dataset = "dcv.2.0")
}
}
