% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-urls.R
\name{flywire_ids}
\alias{flywire_ids}
\title{Flexible specification of flywire ids (including from flytable types)}
\usage{
flywire_ids(
  x,
  file = NULL,
  integer64 = FALSE,
  check_latest = FALSE,
  must_work = FALSE,
  na_ok = FALSE,
  unique = FALSE,
  version = NULL,
  table = c("both", "info", "optic"),
  ...
)
}
\arguments{
\item{x}{A character or bit64::integer64 vector or a dataframe specifying ids
directly \emph{or} a string specifying a query, a URL \emph{or} a
comma/space delimited list of ids (see examples).}

\item{file}{As an alternative to \code{x} the path to a file containing ids.}

\item{integer64}{Whether to return ids as 64 bit integers - more compact than
character vector, but can be more fragile (default \code{FALSE}).}

\item{check_latest}{Whether to check if ids are up to date.}

\item{must_work}{Whether ids must be valid}

\item{na_ok}{whether NA ids are acceptable when \code{must_work=TRUE}}

\item{unique}{Whether to return only unique ids}

\item{version}{Integer materialisation version. The special value of
\code{'latest'} means the most recent materialisation according to CAVE.}

\item{table}{When \code{x} is a query whether to search \code{brain},
\code{optic} lobe or \code{both} info tables.}

\item{...}{Additional arguments passed to \code{\link{flytable_cell_types}}
or \code{\link{ngl_segments}}.}
}
\value{
character (or \code{integer64})) vector of segment ids
}
\description{
allows more flexible specification of flywire root ids compared
  with \code{\link{ngl_segments}} including by queries against cell types
  recorded in flytable. Also useful for reading ids from the clipboard or a
  file, which often consist of a single whitespace or comma-delimited string.
}
\examples{
flywire_ids(data.frame(root_id=1))
flywire_ids(data.frame(root_id=1), integer64=TRUE)
# Bad values will return 0
flywire_ids(data.frame(root_id=-1))
\dontrun{
# will error
flywire_ids(data.frame(root_id=-1), must_work = TRUE)
}
# DL1 olfactory PNs
flywire_ids("DL1_adPN")
# DL1 olfactory PNs but only on the RHS
flywire_ids("DL1_adPN_R")
# specifying materialisation version
flywire_ids("DL1_adPN_R", version=630)
# using SQL wild cards
flywire_ids("DA[12]_\%PN_L")

# all sensory neurons
flywire_ids("super:sensory", integer64=TRUE)

# note that side is defined by soma position (not arbour side)
flywire_ids("class:MBON_R", integer64=TRUE)
# superclass can also have a side specified
flywire_ids("super:motor_R", integer64=TRUE)

# you can also use a comma/whitespace delimited list
flywire_ids("1234, 123456")
# ... which could come from the clipboard
\dontrun{
flywire_ids(clipr::read_clip())

# ... or from a file
flywire_ids(file='~/Downloads/root_ids_Li02_.txt')
}
}
\seealso{
\code{\link{flytable_cell_types}}.

Other neuroglancer-urls: 
\code{\link{flywire_scene}()},
\code{\link{ngl_blank_scene}()},
\code{\link{ngl_decode_scene}()},
\code{\link{ngl_encode_url}()},
\code{\link{ngl_segments}()},
\code{\link{open_fafb_ngl}()}
}
\concept{neuroglancer-urls}
