% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_leaves}
\alias{flywire_leaves}
\title{Find all the supervoxel (leaf) ids that are part of a FlyWire object}
\usage{
flywire_leaves(
  x,
  cloudvolume.url = NULL,
  integer64 = FALSE,
  mip = 0L,
  bbox = NULL,
  ...
)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL chooses the flywire production segmentation
dataset.}

\item{integer64}{Whether to return ids as integer64 type (more compact but a
little fragile) rather than character (default \code{FALSE}).}

\item{mip}{The mip level for the segmentation (expert use only)}

\item{bbox}{The bounding box within which to find supervoxels (default =
\code{NULL} for whole brain. Expert use only.)}

\item{...}{Additional arguments passed to \code{\link{pbsapply}} and
eventually \code{\link{flywire_fetch}} when \code{method="flywire"} OR to
\code{cv$CloudVolume} when \code{method="cloudvolume"}}
}
\description{
Find all the supervoxel (leaf) ids that are part of a FlyWire object
}
\seealso{
\code{\link{flywire_rootid}}
}
