% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosyn.R
\name{flywire_ntplot}
\alias{flywire_ntplot}
\alias{flywire_ntplot3d}
\title{Plot neurotransmitter prediction summaries or synapses in 3D}
\usage{
flywire_ntplot(
  x,
  nts = c("gaba", "acetylcholine", "glutamate", "octopamine", "serotonin", "dopamine"),
  cleft.threshold = 0
)

flywire_ntplot3d(
  x,
  nts = c("gaba", "acetylcholine", "glutamate", "octopamine", "serotonin", "dopamine"),
  plot = c("spheres", "points"),
  cleft.threshold = 0,
  ...
)
}
\arguments{
\item{x}{A flywire rootid or a data.frame of neurotransmitter predictions
returned by \code{\link{flywire_ntpred}}}

\item{nts}{A character vector of neurotransmitters to include in the plot
(default all 6)}

\item{cleft.threshold}{A threshold for the cleft score calculated by Buhmann
et al 2019 (default 0, we have used 30-100 to increase specificity)}

\item{plot}{Whether to plot points or spheres}

\item{...}{additional arguments passed to \code{\link{spheres3d}} or
\code{\link{points3d}}}
}
\description{
\code{flywire_ntplot} plots a ggplot2 histogram of predicted
  neurotransmitter vs prediction probability.

\code{flywire_ntplot3d} makes a 3D plot of synapse location
}
\examples{
\donttest{
# a cholinergic olfactory projection neuron
ntp=flywire_ntpred("720575940615237849")
flywire_ntplot(ntp)
flywire_ntplot(ntp, nts=c("gaba", "acetylcholine", "glutamate"))
flywire_ntplot(ntp, nts=c("gaba", "acetylcholine", "glutamate"), cleft.threshold=100)
}
\dontrun{
flywire_ntplot3d(ntp, nts=c("gaba", "acetylcholine",
  "glutamate"), plot='points', cleft.threshold=30, size=5)
}
}
\seealso{
Other automatic-synapses: 
\code{\link{flywire_ntpred}()},
\code{\link{flywire_partners}()}

Other automatic-synapses: 
\code{\link{flywire_ntpred}()},
\code{\link{flywire_partners}()}
}
\concept{automatic-synapses}
