% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosyn.R
\name{flywire_ntpred}
\alias{flywire_ntpred}
\alias{print.ntprediction}
\title{Return raw neurotransmitter prediction results for output of flywire neuron}
\usage{
flywire_ntpred(x)

\method{print}{ntprediction}(x, ...)
}
\arguments{
\item{x}{A single root id as a string OR a \code{data.frame} of output
(downstream) partners returned by \code{flywire_partners}.}

\item{...}{additional arguments passed to \code{\link{print}}}
}
\value{
A \code{data.frame} of neurotransmitter predictions
}
\description{
the \code{print.ntprediction} method provides a quick summary of
  the neurotransmitter prediction for all output synapses.
}
\examples{
\donttest{
# an olfactory projection neuron
flywire_ntpred("720575940615237849")
# alternatively
\dontrun{
flywire_ntpred(flywire_xyz2id(cbind(116923, 61378, 1474), rawcoords = T))
}
}
}
\seealso{
Other automatic-synapses: 
\code{\link{flywire_ntplot}()},
\code{\link{flywire_partners}()}

Other automatic-synapses: 
\code{\link{flywire_ntplot}()},
\code{\link{flywire_partners}()}
}
\concept{automatic-synapses}
