% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosyn.R
\name{flywire_partners}
\alias{flywire_partners}
\alias{flywire_partner_summary}
\title{Find the input/output partners for a flywire neuron}
\usage{
flywire_partners(
  rootid,
  partners = c("outputs", "inputs", "both"),
  details = FALSE,
  roots = TRUE,
  reference = c("either", "FAFB14", "FlyWire"),
  cloudvolume.url = NULL,
  method = c("auto", "spine", "sqlite"),
  Verbose = TRUE,
  local = NULL,
  ...
)

flywire_partner_summary(
  rootid,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0,
  method = c("auto", "spine", "sqlite"),
  Verbose = NA,
  local = NULL,
  ...
)
}
\arguments{
\item{rootid}{Character vector specifying one or more flywire rootids. As a
convenience for \code{flywire_partner_summary} this argument is passed to
\code{\link{ngl_segments}} allowing you to pass in}

\item{partners}{Whether to fetch input or output synapses or both.}

\item{details}{Whether to include additional details such as X Y Z location
(default \code{FALSE})}

\item{roots}{Whether to fetch the flywire rootids of the partner neurons
(default \code{TRUE})}

\item{reference}{A character vector or a \code{\link{templatebrain}} object
specifying the reference template brain for any 3D coordinate information.
The default value of \code{"either"} will use the natural reference space
of the data source (FAFB14 for sqlite tables, FlyWire for the spine
service).}

\item{cloudvolume.url}{The segmentation source URL for cloudvolume. Normally
you can ignore this and rely on the default segmentation chosen by
\code{\link{choose_segmentation}}}

\item{method}{Whether to use a local SQLite database or remote spine service
for synapse data. The default \code{auto} uses a local database when
available (45GB but faster).}

\item{Verbose}{Whether to print status messages}

\item{local}{path to SQLite synapse data. Evaluated by
\code{fafbseg:::local_or_google}. Work in progress. Default is to download
this data and place it in \code{~/projects/JanFunke}.}

\item{...}{Additional arguments passed to \code{\link{pbsapply}}}

\item{threshold}{For \code{flywire_partner_summary} only return partners with
greater than this number of connections to the query neuron(s) (default of
0 returns all connections)}

\item{remove_autapses}{For \code{flywire_partner_summary} whether to remove
autapses (defaults to TRUE)}

\item{cleft.threshold}{A threshold for the cleft score calculated by Buhmann
et al 2019 (default 0, we have used 30-100 to increase specificity)}
}
\value{
A \code{data.frame} with a \code{regtemplate} attribute specifying
  whether reference brain space for any xyz points.
}
\description{
\code{flywire_partners} is a low level function returning one
  row for every synapse.

\code{flywire_partner_summary} summarises the connectivity of
  one or more flywire neurons.
}
\details{
FIXME behaviour when there are no partners is not well-defined.

  Note that there are many duplicated connections in this raw output, false
  autapses (i.e. the same neuron connected to itself) and other false
  positives. See Buhmann et al for details and ideas about cleaning up the
  results.
}
\examples{
\donttest{
pp=flywire_partners("720575940621039145")
head(pp)
}
\donttest{
flywire_partner_summary("720575940621039145", partners='out')
flywire_partner_summary("720575940621039145", partners='in')
flywire_partner_summary("720575940621039145")

# summary for neuron at a XYZ location (in this case in raw coordinates)
flywire_partner_summary(flywire_xyz2id(cbind(155682, 58180, 3215),
  rawcoords = TRUE))

\dontrun{
# Use Ctrl+Shift+J to share a flywire scene and then do this to get partner
# summary for that URL
flywire_partner_summary(clipr::read_clip())

}
}
}
\seealso{
Other automatic-synapses: 
\code{\link{flywire_adjacency_matrix}()},
\code{\link{flywire_neurons_add_synapses}()},
\code{\link{flywire_ntplot}()},
\code{\link{flywire_ntpred}()}

Other automatic-synapses: 
\code{\link{flywire_adjacency_matrix}()},
\code{\link{flywire_neurons_add_synapses}()},
\code{\link{flywire_ntplot}()},
\code{\link{flywire_ntpred}()}
}
\concept{automatic-synapses}
