% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_rootid}
\alias{flywire_rootid}
\title{Find the root_id of a FlyWire segment / supervoxel.}
\usage{
flywire_rootid(x, ...)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{...}{Additional arguments passed to \code{\link{pbsapply}} (when more
than 1 id) or to \code{\link{flywire_fetch}}}
}
\value{
A vector of root ids as character vectors
}
\description{
Find the root_id of a FlyWire segment / supervoxel.
}
\details{
The main purpose of this function is to convert a supervoxel into
  the current root id for the whole segment. If a segment id has been updated
  due to editing, calling this the original segment id will still return the
  same segment id (although calling it with a supervoxel would return the new
  segment id).
}
\examples{
\dontrun{
flywire_rootid("81489548781649724")
}
}
