% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_rootid}
\alias{flywire_rootid}
\title{Find the root_id of a FlyWire segment / supervoxel.}
\usage{
flywire_rootid(
  x,
  method = c("auto", "cloudvolume", "flywire"),
  cloudvolume.url = NULL,
  ...
)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{method}{Whether to use the flywire API (slow but no python required) OR
cloudvolume (faster for many input ids, but requires python). "auto" (the
default) will choose "flywire" for length 1 queries, "cloudvolume"
otherwise.}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL chooses the flywire production segmentation
dataset.}

\item{...}{Additional arguments passed to \code{\link{pbsapply}} and
eventually \code{\link{flywire_fetch}} when \code{method="flywire"} OR to
\code{cv$CloudVolume} when \code{method="cloudvolume"}}
}
\value{
A vector of root ids as character vectors.
}
\description{
Find the root_id of a FlyWire segment / supervoxel.
}
\details{
The main purpose of this function is to convert a supervoxel into
  the current root id for the whole segment. If a segment id has been updated
  due to editing, calling this the original segment id will still return the
  same segment id (although calling it with a supervoxel would return the new
  segment id).

  There are two \code{method}s. flywire is simpler but will be slower for
  many supervoxels since each id requires a separate http request.
}
\examples{
\donttest{
flywire_rootid(c("81489548781649724", "80011805220634701"))
# same but using the flywire sandbox segmentation
with_segmentation('sandbox', {
flywire_rootid(c("81489548781649724", "80011805220634701"))
})
}
}
