% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-urls.R
\name{flywire_scene}
\alias{flywire_scene}
\title{Return a sample Neuroglancer scene URL for FlyWire dataset}
\usage{
flywire_scene(ids = NULL, open = FALSE)
}
\arguments{
\item{ids}{A set of root ids to include in the scene. Also ccepts a
data.frame containing a column \code{rootid}, \code{flywire.id}, \code{id}
or any form acceptable to \code{\link{ngl_segments}} including neuroglancer
scene URLs.}

\item{open}{Whether to open the scene in your default browser}
}
\value{
A character vector containing a single Neuroglancer URL (invisibly
  when open=TRUE)
}
\description{
Return a sample Neuroglancer scene URL for FlyWire dataset
}
\examples{
\dontrun{
flywire_scene(open=T)
# top 20 partners of a neuron
flywire_scene(flywire_partner_summary("720575940621039145", partners='out')$partner[1:20], open=T)

}
}
\seealso{
Other neuroglancer-urls: 
\code{\link{ngl_blank_scene}()},
\code{\link{ngl_decode_scene}()},
\code{\link{ngl_encode_url}()},
\code{\link{ngl_segments}()},
\code{\link{open_fafb_ngl}()}
}
\concept{neuroglancer-urls}
