% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave.R
\name{flywire_timestamp}
\alias{flywire_timestamp}
\title{Find the timestamp for a given materialisation version}
\usage{
flywire_timestamp(
  version,
  convert = TRUE,
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production")
)
}
\arguments{
\item{version}{Integer materialisation version}

\item{convert}{Whether to convert from Python to R timestamp (default: \code{TRUE})}

\item{datastack_name}{defaults to "flywire_fafb_production". See
\url{https://global.daf-apis.com/info/} for other options.}
}
\value{
A POSIXct object or Python datetime object
}
\description{
Find the timestamp for a given materialisation version
}
\examples{
\donttest{
flywire_timestamp(349)
}
}
\seealso{
Other cave-queries: 
\code{\link{flywire_cave_query}()}
}
\concept{cave-queries}
