% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_updateids}
\alias{flywire_updateids}
\title{Update root ids for flywire neurons using XYZ or supervoxel ids}
\usage{
flywire_updateids(
  x,
  svids = NULL,
  xyz = NULL,
  rawcoords = FALSE,
  voxdims = c(4, 4, 40),
  Verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Current root ids}

\item{svids}{optional supervoxel ids}

\item{xyz}{optional xyz locations in any form understood by
\code{\link{xyzmatrix}}}

\item{rawcoords}{whether the input values are raw voxel indices or in nm}

\item{voxdims}{voxel dimensions in nm used to convert raw coordinates.}

\item{Verbose}{Whether to print a message to the console when updates are
required.}

\item{...}{Additional arguments passed to \code{\link{flywire_islatest}} or
\code{\link{flywire_latestid}}}
}
\value{

}
\description{
Update root ids for flywire neurons using XYZ or supervoxel ids
}
\examples{

kcs=data.frame(
rootid=c("720575940602553568", "720575940602564320", "720575940602605536"),
xyz=c("(159284,42762,3594)", "(159035,41959,3594)", "(157715,44345,3594)")
)
# update root ids
kcs$rootid=flywire_updateids(kcs$rootid, xyz=kcs$xyz)
}
\seealso{
Other flywire-ids: 
\code{\link{flywire_islatest}()},
\code{\link{flywire_latestid}()},
\code{\link{flywire_leaves}()},
\code{\link{flywire_rootid}()},
\code{\link{flywire_xyz2id}()}
}
\concept{flywire-ids}
