% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-user.R
\name{flywire_user_info}
\alias{flywire_user_info}
\title{Fetch information for one or more users}
\usage{
flywire_user_info(
  uids,
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production")
)
}
\arguments{
\item{uids}{A vector of user ids or a dataframe with a \code{user_id} column}

\item{datastack_name}{defaults to the value selected by
\code{\link{choose_segmentation}} and to "flywire_fafb_production" when
that is missing. See \url{https://global.daf-apis.com/info/} for other
options.}
}
\value{
A dataframe of user information
}
\description{
Fetch information for one or more users
}
\examples{
\donttest{
flywire_user_info(60, datastack_name = "flywire_fafb_public")
}
}
