% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-coord.R
\name{flywire_voxdims}
\alias{flywire_voxdims}
\alias{flywire_nm2raw}
\alias{flywire_raw2nm}
\title{Handle raw and nm calibrated flywire coordinates}
\usage{
flywire_voxdims(url = getOption("fafbseg.sampleurl"))

flywire_nm2raw(x, vd = flywire_voxdims())

flywire_raw2nm(x, vd = flywire_voxdims())
}
\arguments{
\item{url}{Optional neuroglancer URL containing voxel size. Defaults to
\code{getOption("fafbseg.sampleurl")} as set by
\code{\link{choose_segmentation}}.}

\item{x}{3D coordinates in any form compatible with \code{\link{xyzmatrix}}}

\item{vd}{The voxel dimensions in nm. Expert use only. Normally found
automatically.}
}
\value{
For \code{flywire_voxdims} A 3-vector

for \code{flywire_raw2nm} and \code{flywire_nm2raw} an Nx3 matrix of
  coordinates
}
\description{
\code{flywire_voxdims} returns the image voxel dimensions which
  are normally used to scale between \bold{raw} and \bold{nm} coordinates.
}
\details{
relies on nat >= 1.10.4
}
\examples{
flywire_voxdims()
# ensure that we use default production flywire scene
with_segmentation('flywire', flywire_voxdims())
flywire_raw2nm(c(159144, 22192, 3560))
flywire_raw2nm('159144 22192 3560')
\dontrun{
flywire_nm2raw(clipr::read_clip())
}
}
