% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_xyz2id}
\alias{flywire_xyz2id}
\title{Title}
\usage{
flywire_xyz2id(
  xyz,
  rawcoords = FALSE,
  voxdims = c(4, 4, 40),
  cloudvolume.url = NULL,
  root = TRUE,
  ...
)
}
\arguments{
\item{xyz}{One or more xyz locations as an Nx3 matrix or in any form
compatible with \code{\link{xyzmatrix}} including \code{neuron} or
\code{mesh3d} surface objects.}

\item{rawcoords}{whether the input values are raw voxel indices or in nm}

\item{voxdims}{voxel dimensions in nm used to convert the}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL choose the production segmentation dataset.}

\item{root}{Whether to return the root id of the whole segment rather than
the supervoxel id.}

\item{...}{additional arguments passed to \code{pbapply} when looking up
multiple positions.}
}
\value{
A character vector of segment ids, \code{NA} when lookup fails.
}
\description{
Title
}
\details{
Note that finding the supervoxel for a given XYZ location is order
  3x faster than finding the root id for the agglomeration of all of the
  super voxels in a given object. Perhaps less intuitively, if you want to
  look up many root ids, it is actually quicker to do
  \code{flywire_xyz2id(,root=F)} followed by \code{\link{flywire_rootid}}
  since that function can look up many root ids in a single call to the
  ChunkedGraph server.
}
\examples{
\donttest{
# sample dataset from FAFB catmaid
n=elmr::dense_core_neurons[[1]]
set.seed(42)
ss=sample(nvertices(n), size=10)
# first convert FAF14 points to FlyWire coordinate space
nx=xform_brain(elmr::dense_core_neurons[[1]], ref="FlyWire", sample="FAFB14")

# now find the ids for the selected xyz locations
flywire_xyz2id(xyzmatrix(nx)[ss,])

# we can also find the supervoxels - much faster
svids=flywire_xyz2id(xyzmatrix(nx)[ss,], root=FALSE)

# now look up the root ids - very fast with method="cloudvolume", the default
flywire_rootid(svids)
}
}
