% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire-api.R
\name{flywire_xyz2id}
\alias{flywire_xyz2id}
\title{Title}
\usage{
flywire_xyz2id(
  xyz,
  rawcoords = FALSE,
  voxdims = c(4, 4, 40),
  cloudvolume.url = NULL,
  root = TRUE,
  ...
)
}
\arguments{
\item{xyz}{One or more xyz locations as an Nx3 matrix or in any form
compatible with \code{\link{xyzmatrix}} including \code{neuron} or
\code{mesh3d} surface objects.}

\item{rawcoords}{whether the input values are raw voxel indices or in nm}

\item{voxdims}{voxel dimensions in nm used to convert the}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL choose the production segmentation dataset.}

\item{root}{Whether to return the root id of the whole segment rather than
the supervoxel id.}

\item{...}{additional arguments passed to \code{pbapply} when looking up
multiple positions.}
}
\value{
A character vector of segment ids, \code{NA} when lookup fails.
}
\description{
Title
}
\examples{
\dontrun{
# sample dataset from FAFB catmaid
n=read.
set.seed(42)
ss=sample(nvertices(n), size=20)
flywire_xyz2id(xyzmatrix(n)[ss,])

# here we actually convert to FlyWire coordinate space which should give a
# much better match
nx=xform_brain(elmr::dense_core_neurons[[1]], ref="FlyWire", sample="FAFB14")
flywire_xyz2id(xyzmatrix(nx)[ss,])
}
}
