% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_decode_scene}
\alias{ngl_decode_scene}
\title{Decode scene information from Neuroglancer URL or JSON block}
\usage{
ngl_decode_scene(x, return.json = FALSE, simplifyVector = TRUE, ...)
}
\arguments{
\item{x}{Character vector containing single Neuroglancer URL or a json block}

\item{return.json}{When \code{TRUE} extracts the JSON block in a URL does not
parse it to an R list}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\description{
Decode scene information from Neuroglancer URL or JSON block
}
\seealso{
\code{\link[utils]{URLdecode}}, \code{\link[jsonlite]{fromJSON}}
\dontrun{
ngl_decode_scene("<someneuroglancerurl>")

# decode scene from URL currently on clipboard
scene=ngl_decode_scene(clipr::read_clip())
}
}
