% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_decode_scene}
\alias{ngl_decode_scene}
\title{Decode scene information from Neuroglancer URL or JSON block}
\usage{
ngl_decode_scene(
  x,
  return.json = FALSE,
  simplifyVector = TRUE,
  simplifyDataFrame = FALSE,
  ...
)
}
\arguments{
\item{x}{Character vector containing single Neuroglancer URL or a json block}

\item{return.json}{When \code{TRUE} extracts the JSON block in a URL does not
parse it to an R list}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{simplifyDataFrame}{coerce JSON arrays containing only records (JSON objects) into a data frame}

\item{...}{additional arguments passed to \code{jsonlite::\link{fromJSON}}}
}
\value{
An R list with additional class \code{ngscene} describing the scene,
  or, when \code{return.json=TRUE}, a character vector.
}
\description{
Decode scene information from Neuroglancer URL or JSON block
}
\examples{

\donttest{
# get sample FlyWire URL
fw_url=with_segmentation('flywire', getOption('fafbseg.sampleurl'))
# only a 0 (dummy) segment id present
ngl_segments(fw_url)
#
fw_sc=ngl_decode_scene(fw_url)
# add a segment
fw_sc$layers[[2]]$segments=union(fw_sc$layers[[2]]$segments,
  "720575940626877799")
# convert back to a URL, nb this depends on choose_segmentation
ngl_encode_url(fw_sc)
\dontrun{
# open in your default browser
browseURL(ngl_encode_url(fw_sc))
}
}

\dontrun{
ngl_decode_scene("<someneuroglancerurl>")

# decode scene from URL currently on clipboard
scene=ngl_decode_scene(clipr::read_clip())

# open a Neuroglancer URL in CATMAID
ngu="<someurl>"
library(elmr)
open_fafb(ngl_decode_scene(ngu))
# Or store the URL rather than opening it
cmu=open_fafb(ngl_decode_scene(ngu), open=FALSE)
}
}
\seealso{
\code{\link[utils]{URLdecode}}, \code{\link[jsonlite]{fromJSON}}

Other neuroglancer-urls: 
\code{\link{ngl_encode_url}()},
\code{\link{open_fafb_ngl}()}
}
\concept{neuroglancer-urls}
