% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_decode_scene}
\alias{ngl_decode_scene}
\title{Decode scene information from Neuroglancer URL or JSON block}
\usage{
ngl_decode_scene(x, return.json = FALSE, simplifyVector = TRUE,
  simplifyDataFrame = FALSE, ...)
}
\arguments{
\item{x}{Character vector containing single Neuroglancer URL or a json block}

\item{return.json}{When \code{TRUE} extracts the JSON block in a URL does not
parse it to an R list}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{simplifyDataFrame}{coerce JSON arrays containing only records (JSON objects) into a data frame}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\value{
An R list with additional class \code{ngscene} describing the scene,
  or, when \code{return.json=TRUE}, a character vector.
}
\description{
Decode scene information from Neuroglancer URL or JSON block
}
\examples{
\dontrun{
ngl_decode_scene("<someneuroglancerurl>")

# decode scene from URL currently on clipboard
scene=ngl_decode_scene(clipr::read_clip())

# open a Neuroglancer URL in CATMAID
ngu="<someurl>"
library(elmr)
open_fafb(ngl_decode_scene(ngu))
# Or store the URL rather than opening it
cmu=open_fafb(ngl_decode_scene(ngu), open=FALSE)
}
}
\seealso{
\code{\link[utils]{URLdecode}}, \code{\link[jsonlite]{fromJSON}}

Other neuroglancer-urls: \code{\link{ngl_encode_url}},
  \code{\link{open_fafb_ngl}}
}
\concept{neuroglancer-urls}
