% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_encode_url}
\alias{ngl_encode_url}
\title{Encode scene information into a neuroglancer URL}
\usage{
ngl_encode_url(body, baseurl = getOption("fafbseg.baseurl"),
  auto_unbox = TRUE, fix_segments = TRUE, ...)
}
\arguments{
\item{body}{A text file or character vector with JSON data or an R list
object}

\item{baseurl}{The base URL including the neuroglancer server}

\item{auto_unbox}{automatically \code{\link{unbox}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link{unbox}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{I()}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}

\item{fix_segments}{Fix URL when only one segment in scene (see details)}

\item{...}{Additional arguments for \code{\link[jsonlite]{toJSON}}}
}
\value{
Character vector containing encoded URL
}
\description{
Encode scene information into a neuroglancer URL
}
\details{
When the neuroglancer URL scene refers to just one segment the only
  way we can currently ensure correctly formed JSON is to add a dummy 0
  segment (thus forming a JSON array).
}
\examples{
\dontrun{
# copy JSON scene information from {} symbol at top right of neuroglancer
# now make a permanent URL for the scene
ngl_encode_url(clipr::read_clip())
}
}
\seealso{
\code{\link{URLencode}}, \code{\link{open_fafb_ngl}},
  \code{\link[jsonlite]{toJSON}}

Other neuroglancer-urls: \code{\link{ngl_decode_scene}},
  \code{\link{open_fafb_ngl}}
}
\concept{neuroglancer-urls}
