% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ngl_encode_url}
\alias{ngl_encode_url}
\title{Encode scene information into a neuroglancer URL}
\usage{
ngl_encode_url(body, baseurl = "https://neuroglancer-demo.appspot.com/#!",
  auto_unbox = TRUE, ...)
}
\arguments{
\item{body}{A text file or character vector with JSON data or an R list
object}

\item{baseurl}{The base URL including the neuroglancer server}

\item{auto_unbox}{automatically \code{\link{unbox}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link{unbox}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{I()}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}

\item{...}{Additional arguments for \code{\link[jsonlite]{toJSON}}}
}
\value{
Character vector containing encoded URL
}
\description{
Encode scene information into a neuroglancer URL
}
\examples{
\dontrun{
# copy JSON scene information from {} symbol at top right of neuroglancer
# now make a permanent URL for the scene
ngl_encode_url(clipr::read_clip())
}
}
\seealso{
\code{\link{URLencode}}, \code{\link{open_fafb_ngl}},
  \code{\link[jsonlite]{toJSON}}
}
