% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ngl_segments}
\alias{ngl_segments}
\title{Helper function to turn diverse inputs into neuroglancer segment ids}
\usage{
ngl_segments(x, as_character = TRUE, include_hidden = FALSE, ...)
}
\arguments{
\item{x}{A neuroglancer scene specification form either in raw JSON format
(character vector), the path to a file on disk, a neuroglancer scene URL
(which embeds a JSON scene specification in a single URL), or an R list
generated by parsing one of the above.}

\item{as_character}{Whether to return segments as character rather than
numeric vector.}

\item{include_hidden}{Whether to include \code{hiddenSegments} (typically for
flywire scenes).}

\item{...}{Additional arguments passed to \code{\link{ngl_decode_scene}}}
}
\value{
Numeric (or character) vector of segment ids
}
\description{
Helper function to turn diverse inputs into neuroglancer segment ids
}
\examples{
# -> character
ngl_segments(c(10950626347, 10952282491, 13307888342))
# turns these into numeric
ngl_segments(c("10950626347", "10952282491", "13307888342"), as_character=FALSE)

\donttest{
u="https://ngl.flywire.ai/?json_url=https://globalv1.flywire-daf.com/nglstate/5409525645443072"
ngl_segments(u, as_character = TRUE)
}

\dontrun{
# from clipboard
ngl_segments(clipr::read_clip())

# URL
ngl_segments("<ngl-scene-url>")
# path to file on disk
ngl_segments("/path/to/scene.json")
# R list
ngl_segments(scenelist)
}
}
