% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{open_fafb_ngl}
\alias{open_fafb_ngl}
\title{Construct Neuroglancer URL based on 3D location data}
\usage{
open_fafb_ngl(x, s = rgl::select3d(), zoomFactor = 8, sampleurl = NULL,
  open = interactive(), ...)
}
\arguments{
\item{x}{A numeric vector OR any object compatible with
\code{\link[nat]{xyzmatrix}} OR a CATMAID URL (see details)}

\item{s}{Optional selection function of the type returned by
\code{\link[rgl]{select3d}}}

\item{zoomFactor}{The Neuroglancer zoomFactor (bigger means zoomed out)}

\item{sampleurl}{A sample URL that defines your neuroglancer dataset.}

\item{open}{Whether or not to open the URL in a browser - this defaults to
\code{TRUE} in interactive use.}

\item{...}{Additional arguments passed to \code{\link[jsonlite]{fromJSON}} to
control JSON parsing.}
}
\description{
Construct Neuroglancer URL based on 3D location data
}
\details{
Neuroglancer scenes seem to be specified in a single URL that
  URLencodes a json object definining layers to display, position etc. This
  function works by taking a sample URL defining such a scene and then
  editing it to point to a new 3D location / adjust zoom. You only need to
  specify the sample URL once per R session. This approach avoids having to
  construct these scene URLs from scratch or embedding the information within
  the package. If you regularly use a particular kind of scene URL, you can
  set \code{options(fafbseg.sampleurl)} in your \code{\link{Rprofile}} file.
}
\examples{
\dontrun{
# Set an existing scene URL (pointing to any old location) to act as
# the template for open_fafb_ngl
options(fafbseg.sampleurl="https://<neuroglancerlurl>")

# Open a location in MB peduncle
open_fafb_ngl(c(433440, 168344, 131200))

# open a CATMAID URL in Neuroglancer
u=paste0("https://fafb.catmaid.virtualflybrain.org/?pid=2&zp=131280&",
"yp=170014.98879622458&xp=426584.81386896875&tool=navigator&sid0=2&s0=-1")
open_fafb_ngl(u)
}
}
