% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton-io.R
\name{read_segments}
\alias{read_segments}
\alias{read_segments2}
\title{Read skeletons for segments by extracting from the corresponding zip file(s)}
\usage{
read_segments(x, voxdims = c(32, 32, 40), ...)
read_segments2(x, voxdims = c(32, 32, 40), minfilesize = 80, ...)
}
\arguments{
\item{x}{A vector of segment ids (you might get these from a neuroglancer
scene)}

\item{voxdims}{The voxel dimensions in nm of the skeletonised data}

\item{...}{additional arguments passed to \code{\link[nat]{read.neurons}}}
}
\value{
A \code{\link[nat]{neuronlist}} containing one
  \code{\link[nat]{neuron}} for each fragment
}
\description{
Read skeletons for segments by extracting from the corresponding zip file(s)

\code{read_segments2} is a reworked version of
  \code{read_segments} that reads skeletons straight from zip files to
  memory.
}
\examples{
\donttest{
n <- read_segments(22427007374)
}
}
\seealso{
\code{\link[nat]{read.neurons}}
}
