% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton-io.R
\name{read_topn}
\alias{read_topn}
\title{Read the largest n segments from a zip file}
\usage{
read_topn(zipfile, n = 1, ...)
}
\description{
Read the largest n segments from a zip file
}
\details{
Note that this will read all the fragments for these segments
}
