% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton-io.R
\name{skelsforsegment}
\alias{skelsforsegment}
\title{Find all skeleton fragments for one segment}
\usage{
skelsforsegment(x)
}
\arguments{
\item{x}{A segment id}
}
\value{
a character vector of fragment names with an attribute containing the
  path to the zip file
}
\description{
Find all skeleton fragments for one segment
}
